@@ -10,7 +10,6 @@ import json
 import logging
 import sys
 import torch
-from helpers.models.smoldit import SmolDiTConfigurationNames
 from helpers.training import quantised_precision_levels
 from helpers.training.optimizer_param import (
     is_optimizer_deprecated,
@@ -18,6 +17,13 @@ from helpers.training.optimizer_param import (
     map_deprecated_optimizer_parameter,
     optimizer_choices,
 )
+from helpers.models.all import (
+    model_families,
+    get_model_flavour_choices,
+    get_all_model_flavours,
+)
+
+model_family_choices = list(model_families.keys())
 
 logger = logging.getLogger("ArgsParser")
 # Are we the primary process?
@@ -86,21 +92,22 @@ def get_argument_parser():
     )
     parser.add_argument(
         "--model_family",
-        choices=[
-            "pixart_sigma",
-            "sana",
-            "kolors",
-            "sd3",
-            "flux",
-            "smoldit",
-            "sdxl",
-            "ltxvideo",
-            "wan",
-            "legacy",
-        ],
+        choices=model_family_choices,
         default=None,
         required=True,
         help=("The model family to train. This option is required."),
+    )
+    parser.add_argument(
+        "--model_flavour",
+        default=None,
+        required=False,
+        choices=get_all_model_flavours(),
+        type=str,
+        help=(
+            "Certain models require designating a given flavour to reference configurations from."
+            " The value for this depends on the model that is selected."
+            f" Currently supported values:{get_model_flavour_choices()}"
+        ),
     )
     parser.add_argument(
         "--model_type",
@@ -108,15 +115,27 @@ def get_argument_parser():
         choices=[
             "full",
             "lora",
-            "deepfloyd-full",
-            "deepfloyd-lora",
-            "deepfloyd-stage2",
-            "deepfloyd-stage2-lora",
         ],
         default="full",
         help=(
             "The training type to use. 'full' will train the full model, while 'lora' will train the LoRA model."
             " LoRA is a smaller model that can be used for faster training."
+        ),
+    )
+    parser.add_argument(
+        "--hidream_use_load_balancing_loss",
+        action="store_true",
+        default=False,
+        help=(
+            "When set, will use the load balancing loss for HiDream training. This is an experimental implementation."
+        ),
+    )
+    parser.add_argument(
+        "--hidream_load_balancing_loss_weight",
+        type=float,
+        default=None,
+        help=(
+            "When set, will use augment the load balancing loss for HiDream training. This is an experimental implementation."
         ),
     )
     parser.add_argument(
@@ -131,6 +150,9 @@ def get_argument_parser():
             "ai-toolkit",
             "tiny",
             "nano",
+            # control / controlnet
+            "all+ffs+embedder",
+            "all+ffs+embedder+controlnet",
         ],
         default="all",
         help=(
@@ -147,12 +169,6 @@ def get_argument_parser():
         ),
     )
     parser.add_argument(
-        "--flow_matching_sigmoid_scale",
-        type=float,
-        default=None,
-        help="Deprecated option. Replaced with --flow_sigmoid_scale.",
-    )
-    parser.add_argument(
         "--flow_sigmoid_scale",
         type=float,
         default=1.0,
@@ -167,12 +183,6 @@ def get_argument_parser():
         ),
     )
     parser.add_argument(
-        "--flux_use_uniform_schedule",
-        default=None,
-        action="store_true",
-        help="Deprecated option. Replaced with --flow_use_uniform_schedule.",
-    )
-    parser.add_argument(
         "--flow_use_uniform_schedule",
         default=False,
         action="store_true",
@@ -180,12 +190,6 @@ def get_argument_parser():
             "Whether or not to use a uniform schedule instead of sigmoid for flow-matching noise schedule."
             " Using uniform sampling may cause a bias toward dark images, and should be used with caution."
         ),
-    )
-    parser.add_argument(
-        "--flux_use_beta_schedule",
-        action="store_true",
-        default=None,
-        help="Deprecated option. Replaced with --flow_use_beta_schedule.",
     )
     parser.add_argument(
         "--flow_use_beta_schedule",
@@ -197,34 +201,16 @@ def get_argument_parser():
         ),
     )
     parser.add_argument(
-        "--flux_beta_schedule_alpha",
-        type=float,
-        default=None,
-        help=("Deprecated option. Replaced with --flux_beta_schedule_alpha."),
-    )
-    parser.add_argument(
         "--flow_beta_schedule_alpha",
         type=float,
         default=2.0,
         help=("The alpha value of the flow-matching beta schedule. Default is 2.0"),
     )
     parser.add_argument(
-        "--flux_beta_schedule_beta",
-        type=float,
-        default=None,
-        help=("Deprecated option. Replaced with --flow_beta_schedule_beta."),
-    )
-    parser.add_argument(
         "--flow_beta_schedule_beta",
         type=float,
         default=2.0,
         help=("The beta value of the flow-matching beta schedule. Default is 2.0"),
-    )
-    parser.add_argument(
-        "--flux_schedule_shift",
-        type=float,
-        default=None,
-        help=("Deprecated option. Replaced with --flow_schedule_shift."),
     )
     parser.add_argument(
         "--flow_schedule_shift",
@@ -237,12 +223,6 @@ def get_argument_parser():
             " details are ignored or accentuated. A higher value will focus more on large compositional features,"
             " and a lower value will focus on the high frequency fine details."
         ),
-    )
-    parser.add_argument(
-        "--flux_schedule_auto_shift",
-        action="store_true",
-        default=None,
-        help="Deprecated option. Replaced with --flow_schedule_auto_shift.",
     )
     parser.add_argument(
         "--flow_schedule_auto_shift",
@@ -346,34 +326,6 @@ def get_argument_parser():
         ),
     )
     parser.add_argument(
-        "--smoldit",
-        action="store_true",
-        default=False,
-        help=("Use the experimental SmolDiT model architecture."),
-    )
-    parser.add_argument(
-        "--smoldit_config",
-        type=str,
-        choices=SmolDiTConfigurationNames,
-        default="smoldit-base",
-        help=(
-            "The SmolDiT configuration to use. This is a list of pre-configured models."
-            " The default is 'smoldit-base'."
-        ),
-    )
-    parser.add_argument(
-        "--flow_matching_loss",
-        type=str,
-        choices=["diffusers", "compatible", "diffusion", "sd35"],
-        default="compatible",
-        help=(
-            "A discrepancy exists between the Diffusers implementation of flow matching and the minimal implementation provided"
-            " by StabilityAI. This experimental option allows switching loss calculations to be compatible with those."
-            " Additionally, 'diffusion' is offered as an option to reparameterise a model to v_prediction loss."
-            " sd35 provides the ability to train on SD3.5's flow-matching target, which is the denoised sample."
-        ),
-    )
-    parser.add_argument(
         "--sd3_clip_uncond_behaviour",
         type=str,
         choices=["empty_string", "zero"],
@@ -462,6 +414,14 @@ def get_argument_parser():
         ),
     )
     parser.add_argument(
+        "--control",
+        action="store_true",
+        default=False,
+        help=(
+            "If set, channel-wise control style training will be used, where a conditioning input image is required alongside the training data."
+        ),
+    )
+    parser.add_argument(
         "--controlnet",
         action="store_true",
         default=False,
@@ -470,8 +430,17 @@ def get_argument_parser():
         ),
     )
     parser.add_argument(
+        "--controlnet_custom_config",
+        type=str,
+        default=None,
+        help=(
+            "When training certain ControlNet models (eg. HiDream) you may set a config containing keys like num_layers or num_single_layers"
+            " to adjust the resulting ControlNet size. This is not supported by most models, and may be ignored if the model does not support it."
+        ),
+    )
+    parser.add_argument(
         "--controlnet_model_name_or_path",
-        action="store_true",
+        type=str,
         default=None,
         help=(
             "When provided alongside --controlnet, this will specify ControlNet model weights to preload from the hub."
@@ -481,7 +450,6 @@ def get_argument_parser():
         "--pretrained_model_name_or_path",
         type=str,
         default=None,
-        required=True,
         help=(
             "Path to pretrained model or model identifier from huggingface.co/models."
             " Some model architectures support loading single-file .safetensors directly."
@@ -534,12 +502,11 @@ def get_argument_parser():
     parser.add_argument(
         "--prediction_type",
         type=str,
-        default="epsilon",
-        choices=["epsilon", "v_prediction", "sample"],
+        default=None,
+        choices=["epsilon", "v_prediction", "sample", "flow_matching"],
         help=(
-            "The type of prediction to use for the u-net. Choose between ['epsilon', 'v_prediction', 'sample']."
-            " For SD 2.1-v, this is v_prediction. For 2.1-base, it is epsilon. SDXL is generally epsilon."
-            " SD 1.5 is epsilon."
+            "For models which support it, you can supply this value to override the prediction type. Choose between ['epsilon', 'v_prediction', 'sample', 'flow_matching']."
+            " This may be needed for some SDXL derivatives that are trained using v_prediction or flow_matching."
         ),
     )
     parser.add_argument(
@@ -640,6 +607,7 @@ def get_argument_parser():
         help=(
             "When using `--timestep_bias_strategy=range`, the final timestep to bias."
             " Defaults to 1000, which is the number of timesteps that SDXL Base and SD 2.x were trained on."
+            " Just to throw a wrench into the works, Kolors was trained on 1100 timesteps."
         ),
     )
     parser.add_argument(
@@ -958,8 +926,7 @@ def get_argument_parser():
         action="store_true",
         help=(
             "When provided, any validation prompt entries in the text embed cache will be recreated."
-            " This is useful if you've modified any of the existing prompts, or, disabled/enabled Compel,"
-            " via `--disable_compel`"
+            " This is useful if you've modified any of the existing prompts."
         ),
     )
     parser.add_argument(
@@ -1067,7 +1034,7 @@ def get_argument_parser():
             "When training diffusion models, the image sizes generally must align to a 64 pixel interval."
             " This is an exception when training models like DeepFloyd that use a base resolution of 64 pixels,"
             " as aligning to 64 pixels would result in a 1:1 or 2:1 aspect ratio, overly distorting images."
-            " For DeepFloyd, this value is set to 8, but all other training defaults to 64. You may experiment"
+            " For DeepFloyd, this value is set to 32, but all other training defaults to 64. You may experiment"
             " with this value, but it is not recommended."
         ),
     )
@@ -1108,15 +1075,18 @@ def get_argument_parser():
         action="store_true",
         help="(SD 2.x only) Whether to train the text encoder. If set, the text encoder should be float32 precision.",
     )
-    # DeepFloyd
     parser.add_argument(
         "--tokenizer_max_length",
         type=int,
         default=None,
         required=False,
-        help="The maximum length of the tokenizer. If not set, will default to the tokenizer's max length.",
+        help=(
+            "The maximum sequence length of the tokenizer output, which defines the sequence length of text embed outputs."
+            " If not set, will default to the tokenizer's max length. Unfortunately, this option only applies to T5 models, and"
+            " due to the biases inducted by sequence length, changing it will result in potentially catastrophic model collapse."
+            " This option causes poor training results. This is normal, and can be expected from changing this value."
+        ),
     )
-    # End DeepFloyd-specific settings
     parser.add_argument(
         "--train_batch_size",
         type=int,
@@ -1152,15 +1122,45 @@ def get_argument_parser():
         ),
     )
     parser.add_argument(
+        "--checkpointing_rolling_steps",
+        type=int,
+        default=0,
+        help=(
+            "Functions similarly to --checkpointing_steps, but only a single rolling checkpoint is ever saved and this checkpoint does not count"
+            " towards the value of --checkpoints_total_limit. Useful for when the underlying runtime environment may be prone to spontaneous"
+            " interruption (e.g. spot instances, unreliable hardware, etc) and saving state more frequently is beneficial. This allows one to"
+            " save normal checkpoints at a reasonable cadence but save a rolling checkpoint more frequently so as to avoid losing progress."
+        ),
+    )
+    parser.add_argument(
+        "--checkpointing_use_tempdir",
+        action="store_true",
+        default=False,
+        help=(
+            "Write saved checkpoint directories to a temporary name and atomically rename after successfully writing all state. This ensures that a"
+            " given checkpoint will never be considered for resuming if it wasn't fully written out - as the state cannot be guaranteed. Useful for"
+            " when the underlying runtime environment may be prone to spontaneous interruption (e.g. spot instances, unreliable hardware, etc)."
+        ),
+    )
+    parser.add_argument(
         "--checkpoints_total_limit",
         type=int,
         default=None,
         help="Max number of checkpoints to store.",
     )
     parser.add_argument(
+        "--checkpoints_rolling_total_limit",
+        type=int,
+        default=1,
+        help=(
+            "Max number of rolling checkpoints to store. One almost always wants this to be 1, but there could be a usecase where one desires to"
+            " run a shorter window of more frequent checkpoints alongside a normal checkpointing interval done at less frequent steps."
+        ),
+    )
+    parser.add_argument(
         "--resume_from_checkpoint",
         type=str,
-        default=None,
+        default="latest",
         help=(
             "Whether training should be resumed from a previous checkpoint. Use a path saved by"
             ' `--checkpointing_steps`, or `"latest"` to automatically select the last available checkpoint.'
@@ -1208,6 +1208,12 @@ def get_argument_parser():
         help="Scale the learning rate by the number of GPUs, gradient accumulation steps, and batch size.",
     )
     parser.add_argument(
+        "--lr_scale_sqrt",
+        action="store_true",
+        default=False,
+        help="If using --lr-scale, use the square root of (number of GPUs * gradient accumulation steps * batch size).",
+    )
+    parser.add_argument(
         "--lr_scheduler",
         type=str,
         default="sine",
@@ -1240,6 +1246,25 @@ def get_argument_parser():
         default=0.8,
         help="Power factor of the polynomial scheduler.",
     )
+    parser.add_argument(
+        "--distillation_method",
+        default=None,
+        choices=["dcm"],
+        help=(
+            "The distillation method to use. Currently, only 'dcm' is supported via LoRA."
+            " This will apply the selected distillation method to the model."
+        ),
+    )
+    parser.add_argument(
+        "--distillation_config",
+        default=None,
+        type=str,
+        help=(
+            "The config for your selected distillation method."
+            " If passing it via config.json, simply provide the JSON object directly."
+        ),
+    )
+
     parser.add_argument(
         "--use_ema",
         action="store_true",
@@ -1310,6 +1335,14 @@ def get_argument_parser():
         help=(
             "Revision of pretrained non-ema model identifier. Must be a branch, tag or git identifier of the local or"
             " remote repository specified with --pretrained_model_name_or_path."
+        ),
+    )
+    parser.add_argument(
+        "--offload_during_startup",
+        action="store_true",
+        help=(
+            "When set, text encoders, the VAE, or other models will be moved to and from the CPU as needed, which can slow"
+            " down startup, but saves VRAM. This is useful for video models or high-resolution pre-caching of latent embeds."
         ),
     )
     parser.add_argument(
@@ -1480,14 +1513,6 @@ def get_argument_parser():
         ),
     )
     parser.add_argument(
-        "--benchmark_base_model",
-        action="store_true",
-        default=False,
-        help=(
-            "Deprecated option, benchmarks are now enabled by default. Use --disable_benchmark to disable."
-        ),
-    )
-    parser.add_argument(
         "--disable_benchmark",
         action="store_true",
         default=False,
@@ -1626,13 +1651,6 @@ def get_argument_parser():
         help=(
             "When generating embeds for Sana, a complex human instruction will be attached to your prompt by default."
             " This is required for the Gemma model to produce meaningful image caption embeds."
-        ),
-    )
-    parser.add_argument(
-        "--allow_tf32",
-        action="store_true",
-        help=(
-            "Deprecated option. TF32 is now enabled by default. Use --disable_tf32 to disable."
         ),
     )
     parser.add_argument(
@@ -1817,12 +1835,12 @@ def get_argument_parser():
     parser.add_argument(
         "--validation_noise_scheduler",
         type=str,
-        choices=["ddim", "ddpm", "euler", "euler-a", "unipc"],
+        choices=["ddim", "ddpm", "euler", "euler-a", "unipc", "dpm++"],
         default=None,
         help=(
             "When validating the model at inference time, a different scheduler may be chosen."
             " UniPC can offer better speed, and Euler A can put up with instabilities a bit better."
-            " For zero-terminal SNR models, DDIM is the best choice. Choices: ['ddim', 'ddpm', 'euler', 'euler-a', 'unipc'],"
+            " For zero-terminal SNR models, DDIM is the best choice. Choices: ['ddim', 'ddpm', 'euler', 'euler-a', 'unipc', 'dpm++'],"
             " Default: None (use the model default)"
         ),
     )
@@ -1913,7 +1931,7 @@ def get_argument_parser():
             " and may not be necessary for your use case."
         ),
     )
-    for i in range(1, 4):
+    for i in range(1, 5):
         parser.add_argument(
             f"--text_encoder_{i}_precision",
             type=str,
@@ -1966,11 +1984,6 @@ def get_argument_parser():
         ),
     )
     parser.add_argument(
-        "--enable_xformers_memory_efficient_attention",
-        action="store_true",
-        help="Whether or not to use xformers. Deprecated and slated for future removal. Use --attention_mechanism.",
-    )
-    parser.add_argument(
         "--set_grads_to_none",
         action="store_true",
         help=(
@@ -1995,6 +2008,11 @@ def get_argument_parser():
         ),
     )
     parser.add_argument(
+        "--masked_loss_probability",
+        type=float,
+        default=1.0,
+    )
+    parser.add_argument(
         "--validation_guidance",
         type=float,
         default=7.5,
@@ -2004,7 +2022,7 @@ def get_argument_parser():
         "--validation_guidance_real",
         type=float,
         default=1.0,
-        help="Use real CFG sampling for Flux validation images. Default: 1.0 (no CFG)",
+        help="Use real CFG sampling for distilled models. Default: 1.0 (no CFG)",
     )
     parser.add_argument(
         "--validation_no_cfg_until_timestep",
@@ -2300,6 +2318,21 @@ def parse_cmdline_args(input_args=None, exit_on_error: bool = False):
     if args is None and exit_on_error:
         sys.exit(1)
 
+    if (
+        args.controlnet_custom_config is not None
+        and type(args.controlnet_custom_config) is str
+    ):
+        if args.controlnet_custom_config.startswith("{"):
+            try:
+                import ast
+
+                args.controlnet_custom_config = ast.literal_eval(
+                    args.controlnet_custom_config
+                )
+            except Exception as e:
+                logger.error(f"Could not load controlnet_custom_config: {e}")
+                raise
+
     if args.optimizer == "adam_bfloat16" and args.mixed_precision != "bf16":
         if not torch.backends.mps.is_available():
             logging.error(
@@ -2404,7 +2437,7 @@ def parse_cmdline_args(input_args=None, exit_on_error: bool = False):
             warning_log(
                 "MPS may benefit from the use of --unet_attention_slice for memory savings at the cost of speed."
             )
-        if args.model_family != "smoldit" and args.train_batch_size > 16:
+        if args.train_batch_size > 16:
             error_log(
                 "An M3 Max 128G will use 12 seconds per step at a batch size of 1 and 65 seconds per step at a batch size of 12."
                 " Any higher values will result in NDArray size errors or other unstable training results and crashes."
@@ -2451,16 +2484,6 @@ def parse_cmdline_args(input_args=None, exit_on_error: bool = False):
         )
         info_log(f"Default VAE Cache location: {args.cache_dir_vae}")
         info_log(f"Text Cache location: {args.cache_dir_text}")
-    if args.model_family == "sd3":
-        warning_log(
-            "MM-DiT requires an alignment value of 64px. Overriding the value of --aspect_bucket_alignment."
-        )
-        args.aspect_bucket_alignment = 64
-        if args.sd3_t5_uncond_behaviour is None:
-            args.sd3_t5_uncond_behaviour = args.sd3_clip_uncond_behaviour
-        info_log(
-            f"SD3 embeds for unconditional captions: t5={args.sd3_t5_uncond_behaviour}, clip={args.sd3_clip_uncond_behaviour}"
-        )
 
     elif "deepfloyd" in args.model_type:
         deepfloyd_pixel_alignment = 8
@@ -2510,93 +2533,14 @@ def parse_cmdline_args(input_args=None, exit_on_error: bool = False):
     if args.timestep_bias_portion < 0.0 or args.timestep_bias_portion > 1.0:
         raise ValueError("Timestep bias portion must be between 0.0 and 1.0.")
 
-    if args.controlnet and "lora" in args.model_type:
-        raise ValueError("ControlNet is not supported for LoRA models.")
-
     if args.metadata_update_interval < 60:
         raise ValueError("Metadata update interval must be at least 60 seconds.")
 
-    if args.model_family == "sd3":
-        args.pretrained_vae_model_name_or_path = None
-        args.disable_compel = True
-
-    t5_max_length = 154
-    if args.model_family == "sd3" and (
-        args.tokenizer_max_length is None
-        or int(args.tokenizer_max_length) > t5_max_length
-    ):
-        if not args.i_know_what_i_am_doing:
-            warning_log(
-                f"Updating T5 XXL tokeniser max length to {t5_max_length} for SD3."
-            )
-            args.tokenizer_max_length = t5_max_length
-        else:
-            warning_log(
-                f"-!- SD3 supports a max length of {t5_max_length} tokens, but you have supplied `--i_know_what_i_am_doing`, so this limit will not be enforced. -!-"
-            )
-            warning_log(
-                f"The model will begin to collapse after a short period of time, if the model you are continuing from has not been tuned beyond {t5_max_length} tokens."
-            )
-    flux_version = "dev"
-    model_max_seq_length = 512
-    if (
-        "schnell" in args.pretrained_model_name_or_path.lower()
-        or args.flux_fast_schedule
-    ):
-        if not args.flux_fast_schedule and not args.i_know_what_i_am_doing:
-            error_log(
-                "Schnell requires --flux_fast_schedule (or --i_know_what_i_am_doing)."
-            )
-            sys.exit(1)
-        flux_version = "schnell"
-        model_max_seq_length = 256
-
-    if args.model_family == "wan":
-        args.tokenizer_max_length = 226
-    if args.model_family in ["wan", "ltxvideo"]:
-        info_log("Disabling unconditional validation to save on time.")
-        args.validation_disable_unconditional = True
-
-    if args.model_family == "flux":
-        if (
-            args.tokenizer_max_length is None
-            or int(args.tokenizer_max_length) > model_max_seq_length
-        ):
-            if not args.i_know_what_i_am_doing:
-                warning_log(
-                    f"Updating T5 XXL tokeniser max length to {model_max_seq_length} for Flux."
-                )
-                args.tokenizer_max_length = model_max_seq_length
-            else:
-                warning_log(
-                    f"-!- Flux supports a max length of {model_max_seq_length} tokens, but you have supplied `--i_know_what_i_am_doing`, so this limit will not be enforced. -!-"
-                )
-                warning_log(
-                    f"The model will begin to collapse after a short period of time, if the model you are continuing from has not been tuned beyond 256 tokens."
-                )
-        if flux_version == "dev":
-            if args.validation_num_inference_steps > 28:
-                warning_log(
-                    "Flux Dev expects around 28 or fewer inference steps. Consider limiting --validation_num_inference_steps to 28."
-                )
-            if args.validation_num_inference_steps < 15:
-                warning_log(
-                    "Flux Dev expects around 15 or more inference steps. Consider increasing --validation_num_inference_steps to 15."
-                )
-        if flux_version == "schnell" and args.validation_num_inference_steps > 4:
-            warning_log(
-                "Flux Schnell requires fewer inference steps. Consider reducing --validation_num_inference_steps to 4."
-            )
-
-        if args.flux_guidance_mode == "mobius":
-            warning_log(
-                "Mobius training is only for the most elite. Pardon my English, but this is not for those who don't like to destroy something beautiful every now and then. If you feel perhaps this is not for you, please consider using a different guidance mode."
-            )
-            if args.flux_guidance_min < 1.0:
-                warning_log(
-                    "Flux minimum guidance value for Mobius training is 1.0. Updating value.."
-                )
-                args.flux_guidance_min = 1.0
+    args.vae_path = (
+        args.pretrained_model_name_or_path
+        if args.pretrained_vae_model_name_or_path is None
+        else args.pretrained_vae_model_name_or_path
+    )
 
     if args.use_ema and args.ema_cpu_only:
         args.ema_device = "cpu"
@@ -2607,13 +2551,6 @@ def parse_cmdline_args(input_args=None, exit_on_error: bool = False):
         error_log("Both --optimizer_beta1 and --optimizer_beta2 should be provided.")
         sys.exit(1)
 
-    if not args.i_know_what_i_am_doing:
-        if args.model_family == "pixart_sigma":
-            if args.max_grad_norm is None or float(args.max_grad_norm) > 0.01:
-                warning_log(
-                    f"PixArt Sigma requires --max_grad_norm=0.01 to prevent model collapse. Overriding value. Set this value manually to disable this warning."
-                )
-                args.max_grad_norm = 0.01
     if args.gradient_checkpointing:
         # enable torch compile w/ activation checkpointing :[ slows us down.
         torch._dynamo.config.optimize_ddp = False
@@ -2718,6 +2655,16 @@ def parse_cmdline_args(input_args=None, exit_on_error: bool = False):
                 )
                 args.lora_initialisation_style = "default"
 
+    if args.distillation_config is not None:
+        if args.distillation_config.startswith("{"):
+            try:
+                import ast
+
+                args.distillation_config = ast.literal_eval(args.distillation_config)
+            except Exception as e:
+                logger.error(f"Could not load distillation_config: {e}")
+                raise
+
     if not args.data_backend_config:
         from helpers.training.state_tracker import StateTracker
 
@@ -2740,13 +2687,6 @@ def parse_cmdline_args(input_args=None, exit_on_error: bool = False):
             f"Invalid gradient_accumulation_steps parameter: {args.gradient_accumulation_steps}, should be >= 1"
         )
 
-    if args.base_model_precision == 'fp8-quanto':
-        if args.model_family in ['wan', 'sd3']:
-            error_log(
-                "Quanto does not support WAN or SD3 models for FP8. Please use torchao for FP8, or int8 instead."
-            )
-            sys.exit(1)
-
     if args.validation_guidance_skip_layers is not None:
         if args.model_family not in ["sd3", "wan"]:
             raise ValueError(
@@ -2763,14 +2703,6 @@ def parse_cmdline_args(input_args=None, exit_on_error: bool = False):
         except Exception as e:
             logger.error(f"Could not load validation_guidance_skip_layers: {e}")
             raise
-
-    if args.framerate is None:
-        if args.model_family == "ltxvideo":
-            args.framerate = 25
-        if args.model_family == "wan":
-            args.framerate = 15
-            args.vae_enable_tiling = True
-            args.vae_enable_slicing = True
 
     if (
         args.sana_complex_human_instruction is not None
@@ -2791,26 +2723,14 @@ def parse_cmdline_args(input_args=None, exit_on_error: bool = False):
     elif args.sana_complex_human_instruction == "None":
         args.sana_complex_human_instruction = None
 
-    if args.enable_xformers_memory_efficient_attention:
-        if args.attention_mechanism != "xformers":
-            warning_log(
-                "The option --enable_xformers_memory_efficient_attention is deprecated. Please use --attention_mechanism=xformers instead."
-            )
-            args.attention_mechanism = "xformers"
-
     if args.attention_mechanism != "diffusers" and not torch.cuda.is_available():
         warning_log(
             "For non-CUDA systems, only Diffusers attention mechanism is officially supported."
         )
 
     deprecated_options = {
-        "flux_beta_schedule_alpha": "flow_beta_schedule_alpha",
-        "flux_beta_schedule_beta": "flow_beta_schedule_beta",
-        "flux_use_beta_schedule": "flow_use_beta_schedule",
-        "flux_use_uniform_schedule": "flow_use_uniform_schedule",
-        "flux_schedule_shift": "flow_schedule_shift",
-        "flux_schedule_auto_shift": "flow_schedule_auto_shift",
-        "flow_matching_sigmoid_scale": "flow_sigmoid_scale",
+        # how to deprecate options:
+        # "flux_beta_schedule_alpha": "flow_beta_schedule_alpha",
     }
 
     for deprecated_option, replacement_option in deprecated_options.items():
