@@ -441,7 +441,11 @@ class ParquetMetadataBackend(MetadataBackend):
         if isinstance(series_or_scalar, pd.Series):
             series_or_scalar = series_or_scalar.iloc[0]  # Just unwrap the first value
         elif isinstance(series_or_scalar, str):
-            series_or_scalar = float(series_or_scalar) if '.' in series_or_scalar else int(series_or_scalar)
+            series_or_scalar = (
+                float(series_or_scalar)
+                if "." in series_or_scalar
+                else int(series_or_scalar)
+            )
 
         # After unwrapping, if it's an np.int* or np.float*, cast to python int/float
         if isinstance(series_or_scalar, np.integer):
@@ -454,7 +458,6 @@ class ParquetMetadataBackend(MetadataBackend):
             return None
         else:
             raise ValueError(f"Unsupported data type: {type(series_or_scalar)}")
-
 
     def _process_for_bucket(
         self,
