@@ -8,56 +8,6 @@ from diffusers.pipelines.flux.pipeline_flux import (
 )
 
 
-def apply_flow_schedule_shift(args, noise_scheduler, sigmas, noise):
-    # Resolution-dependent shifting of timestep schedules as per section 5.3.2 of SD3 paper
-    shift = None
-    if args.flow_schedule_shift is not None and args.flow_schedule_shift > 0:
-        # Static shift value for every resolution
-        shift = args.flow_schedule_shift
-    elif args.flow_schedule_auto_shift:
-        # Resolution-dependent shift value calculation used by official Flux inference implementation
-        image_seq_len = (noise.shape[-1] * noise.shape[-2]) // 4
-        mu = calculate_shift_flux(
-            (noise.shape[-1] * noise.shape[-2]) // 4,
-            noise_scheduler.config.base_image_seq_len,
-            noise_scheduler.config.max_image_seq_len,
-            noise_scheduler.config.base_shift,
-            noise_scheduler.config.max_shift,
-        )
-        shift = math.exp(mu)
-    if shift is not None:
-        sigmas = (sigmas * shift) / (1 + (shift - 1) * sigmas)
-    return sigmas
-
-
-def get_mobius_guidance(args, global_step, steps_per_epoch, batch_size, device):
-    """
-    state of the art
-    """
-    steps_remaining = steps_remaining_in_epoch(global_step, steps_per_epoch)
-
-    # Start with a linear mapping from remaining steps to a scale between 0 and 1
-    scale_factor = steps_remaining / steps_per_epoch
-
-    # we want the last 10% of the epoch to have a guidance of 1.0
-    threshold_step_count = max(1, int(steps_per_epoch * 0.1))
-
-    if (
-        steps_remaining <= threshold_step_count
-    ):  # Last few steps in the epoch, set guidance to 1.0
-        guidance_values = [1.0 for _ in range(batch_size)]
-    else:
-        # Sample between flux_guidance_min and flux_guidance_max with bias towards 1.0
-        guidance_values = [
-            random.uniform(args.flux_guidance_min, args.flux_guidance_max)
-            * scale_factor
-            + (1.0 - scale_factor)
-            for _ in range(batch_size)
-        ]
-
-    return guidance_values
-
-
 def update_flux_schedule_to_fast(args, noise_scheduler_to_copy):
     if args.flux_fast_schedule and args.model_family.lower() == "flux":
         # 4-step noise schedule [0.7, 0.1, 0.1, 0.1] from SD3-Turbo paper
@@ -119,3 +69,37 @@ def prepare_latent_image_ids(batch_size, height, width, device, dtype):
     )
 
     return latent_image_ids.to(device=device, dtype=dtype)[0]
+
+
+def build_kontext_inputs(
+    cond_latents: torch.Tensor,
+    dtype: torch.dtype,
+    device: torch.device,
+    latent_channels: int,
+):
+    """
+    Args
+    ----
+    cond_latents : (B, C, H, W)  – already VAE-encoded by VAECache
+    dtype        : dtype to use (match main latents)
+    device       : target device
+    latent_channels : self.latent_channels (16 for Flux)
+
+    Returns
+    -------
+    packed_cond : (B, S, C*4)   – flattened patch sequence
+    cond_ids    : (B, S, 3)     – seq-ids with id[...,0] == 1
+    """
+    B, C, H, W = cond_latents.shape  # (C should match latent_channels)
+    packed_cond = pack_latents(cond_latents, B, C, H, W).to(device=device, dtype=dtype)
+
+    # seq-ids: flag-channel==1, rest is y/x indices
+    idx_y = torch.arange(H // 2, device=device)
+    idx_x = torch.arange(W // 2, device=device)
+    ids = torch.stack(
+        torch.meshgrid(idx_y, idx_x, indexing="ij"), dim=-1
+    )  # (H/2,W/2,2)
+    ones = torch.ones_like(ids[..., :1])
+    ids = torch.cat([ones, ids], dim=-1).view(1, -1, 3).expand(B, -1, -1).to(dtype)
+
+    return packed_cond, ids
