@@ -27,6 +27,7 @@ class MultiAspectSampler(torch.utils.data.Sampler):
         id: str,
         metadata_backend: MetadataBackend,
         data_backend: BaseDataBackend,
+        model,
         accelerator,
         batch_size: int,
         debug_aspect_buckets: bool = False,
@@ -58,6 +59,7 @@ class MultiAspectSampler(torch.utils.data.Sampler):
             raise ValueError(
                 f"Sampler ID ({self.id}) must match DataBackend ID ({data_backend.id}) and MetadataBackend ID ({metadata_backend.id})."
             )
+        self.model = model
         # Update the logger name with the id:
         self.dataset_type = dataset_type
         self.sample_type_str = "image"
@@ -69,17 +71,26 @@ class MultiAspectSampler(torch.utils.data.Sampler):
             f"MultiAspectSampler-{self.id}",
             os.environ.get("SIMPLETUNER_LOG_LEVEL", "INFO"),
         )
-        if conditioning_type is not None:
-            if conditioning_type not in ["controlnet", "mask"]:
-                raise ValueError(
-                    f"Unknown conditioning image type: {conditioning_type}"
-                )
+
+        self._val_cursor = 0
+        self._val_master_list = []
         self.conditioning_type = conditioning_type
         self.is_regularisation_data = is_regularisation_data
 
         self.rank_info = rank_info()
         self.accelerator = accelerator
         self.metadata_backend = metadata_backend
+        if conditioning_type is not None:
+            if conditioning_type not in [
+                "controlnet",
+                "mask",
+                "segmentation",
+                "reference_strict",
+                "reference_loose",
+            ]:
+                raise ValueError(
+                    f"Unknown conditioning image type: {conditioning_type}"
+                )
         self.data_backend = data_backend
         self.current_bucket = None
         self.current_epoch = 1
@@ -97,6 +108,9 @@ class MultiAspectSampler(torch.utils.data.Sampler):
         self.exhausted_buckets = []
         self.buckets = self.load_buckets()
         self.state_manager = BucketStateManager(self.id)
+        self._val_master_list = sorted(
+            sum(self.metadata_backend.aspect_ratio_bucket_indices.values(), [])
+        )
 
     def save_state(self, state_path: str):
         """
@@ -157,7 +171,7 @@ class MultiAspectSampler(torch.utils.data.Sampler):
             []
         )  # [tuple(validation_shortname, validation_prompt, validation_sample)]
         for img_idx in range(batch_size):
-            image_path = self._yield_random_image()
+            image_path = self._yield_sequential_image()
             image_data = self.data_backend.read_image(image_path)
             image_metadata = self.metadata_backend.get_metadata_by_filepath(image_path)
             training_sample = TrainingSample(
@@ -165,6 +179,7 @@ class MultiAspectSampler(torch.utils.data.Sampler):
                 data_backend_id=self.id,
                 image_metadata=image_metadata,
                 image_path=image_path,
+                model=self.model,
             )
             training_sample.prepare()
             validation_shortname = f"{self.id}_{img_idx}"
@@ -178,12 +193,17 @@ class MultiAspectSampler(torch.utils.data.Sampler):
                 instance_prompt=self.instance_prompt,
             )
             if type(validation_prompt) == list:
-                validation_prompt = random.choice(validation_prompt)
                 self.debug_log(
                     f"Selecting random prompt from list: {validation_prompt}"
                 )
+                validation_prompt = random.choice(validation_prompt)
             results.append(
-                (validation_shortname, validation_prompt, training_sample.image)
+                (
+                    validation_shortname,
+                    validation_prompt,
+                    image_path,
+                    training_sample.image,
+                )
             )
 
         return results
@@ -207,6 +227,22 @@ class MultiAspectSampler(torch.utils.data.Sampler):
 
         to_yield = self._validate_and_yield_images_from_samples(samples, bucket)
         return to_yield
+
+    def _yield_sequential_image(self):
+        """
+        Always return the *next* image in a fixed list, wrapping around when
+        we reach the end.  Nothing is random, so the same N calls → same N
+        paths every run.
+        """
+        if self._val_cursor >= len(self._val_master_list):
+            self._val_cursor = 0
+            if len(self._val_master_list) == 0:
+                raise MultiDatasetExhausted(
+                    "No validation images available. Please check your dataset."
+                )
+        path = self._val_master_list[self._val_cursor]
+        self._val_cursor = (self._val_cursor + 1) % len(self._val_master_list)
+        return path
 
     def _yield_random_image(self):
         bucket = random.choice(self.buckets)
@@ -422,13 +458,11 @@ class MultiAspectSampler(torch.utils.data.Sampler):
             if image_metadata is None:
                 image_metadata = {}
             if (
-                StateTracker.get_args().model_type
+                StateTracker.get_args().model_family
                 not in [
-                    "legacy",
-                    "deepfloyd-full",
-                    "deepfloyd-lora",
-                    "deepfloyd-stage2",
-                    "deepfloyd-stage2-lora",
+                    "sd1x",
+                    "sd2x",
+                    "deepfloyd",
                 ]
                 and "crop_coordinates" not in image_metadata
             ):
@@ -465,9 +499,12 @@ class MultiAspectSampler(torch.utils.data.Sampler):
         """
         # strip leading /
         original_sample_path = original_sample_path.lstrip("/")
-        full_path = os.path.join(
-            self.metadata_backend.instance_data_dir, original_sample_path
-        )
+        if self.metadata_backend.instance_data_dir not in original_sample_path:
+            full_path = os.path.join(
+                self.metadata_backend.instance_data_dir, original_sample_path
+            )
+        else:
+            full_path = original_sample_path
         try:
             conditioning_sample_data = self.data_backend.read_image(full_path)
         except Exception as e:
@@ -484,6 +521,7 @@ class MultiAspectSampler(torch.utils.data.Sampler):
             image_metadata=self.metadata_backend.get_metadata_by_filepath(full_path),
             image_path=full_path,
             conditioning_type=self.conditioning_type,
+            model=self.model,
         )
         return conditioning_sample
 
@@ -495,9 +533,14 @@ class MultiAspectSampler(torch.utils.data.Sampler):
         sampler = conditioning_dataset["sampler"]
         outputs = list(samples)
         for sample in samples:
-            sample_path = sample["image_path"].split(
-                self.metadata_backend.instance_data_dir
-            )[-1]
+            sample_path = sample["image_path"]
+            if (
+                self.metadata_backend.instance_data_dir is not None
+                and self.metadata_backend.instance_data_dir != ""
+            ):
+                sample_path = sample_path.split(
+                    self.metadata_backend.instance_data_dir
+                )[-1]
             conditioning_sample = sampler.get_conditioning_sample(sample_path)
             outputs.append(conditioning_sample)
         return tuple(outputs)
