@@ -3,32 +3,36 @@ import diffusers
 import os
 import wandb
 import logging
+import inspect
 import sys
 import numpy as np
 from tqdm import tqdm
 from helpers.training.wrappers import unwrap_model
+from helpers.models.common import VideoModelFoundation, ImageModelFoundation
+from helpers.models.common import ModelFoundation
+
+try:
+    import pillow_jxl
+except ModuleNotFoundError:
+    pass
 from PIL import Image
 from helpers.training.state_tracker import StateTracker
-from helpers.data_backend.factory import move_text_encoders
+from helpers.models.common import PredictionTypes, PipelineTypes
 from helpers.training.exceptions import MultiDatasetExhausted
-from helpers.models.sdxl.pipeline import (
-    StableDiffusionXLPipeline,
-    StableDiffusionXLImg2ImgPipeline,
-)
-from helpers.legacy.pipeline import StableDiffusionPipeline
 from diffusers.schedulers import (
     EulerDiscreteScheduler,
     EulerAncestralDiscreteScheduler,
     FlowMatchEulerDiscreteScheduler,
     UniPCMultistepScheduler,
+    DPMSolverMultistepScheduler,
     DDIMScheduler,
     DDPMScheduler,
 )
+from helpers.models.hidream.schedule import FlowUniPCMultistepScheduler
 from diffusers.utils.torch_utils import is_compiled_module
 from helpers.multiaspect.image import MultiaspectImage
 from helpers.image_manipulation.brightness import calculate_luminance
 from PIL import Image, ImageDraw, ImageFont
-from diffusers import SanaPipeline
 from helpers.training.deepspeed import (
     deepspeed_zero_init_disabled_context_manager,
     prepare_model_for_deepspeed,
@@ -38,24 +42,16 @@ from transformers.utils import ContextManagers
 logger = logging.getLogger(__name__)
 logger.setLevel(os.environ.get("SIMPLETUNER_LOG_LEVEL") or "INFO")
 
-try:
-    from helpers.models.sd3.pipeline import (
-        StableDiffusion3Pipeline,
-        StableDiffusion3Img2ImgPipeline,
-    )
-except ImportError:
-    logger.error(
-        "Stable Diffusion 3 not available in this release of Diffusers. Please upgrade."
-    )
-    raise ImportError()
 
 SCHEDULER_NAME_MAP = {
     "euler": EulerDiscreteScheduler,
     "euler-a": EulerAncestralDiscreteScheduler,
-    "flow-match": FlowMatchEulerDiscreteScheduler,
+    "flow_matching": FlowMatchEulerDiscreteScheduler,
     "unipc": UniPCMultistepScheduler,
+    "flow_unipc": FlowUniPCMultistepScheduler,
     "ddim": DDIMScheduler,
     "ddpm": DDPMScheduler,
+    "dpm++": DPMSolverMultistepScheduler,
     "sana": FlowMatchEulerDiscreteScheduler,
 }
 
@@ -81,7 +77,7 @@ def resize_validation_images(validation_images, edge_length):
     # we have to scale all the inputs to a stage4 image down to 64px smaller edge.
     resized_validation_samples = []
     for _sample in validation_images:
-        validation_shortname, validation_prompt, training_sample_image = _sample
+        validation_shortname, validation_prompt, _, training_sample_image = _sample
         resize_to, crop_to, new_aspect_ratio = (
             MultiaspectImage.calculate_new_size_by_pixel_edge(
                 aspect_ratio=MultiaspectImage.calculate_image_aspect_ratio(
@@ -173,9 +169,23 @@ def retrieve_validation_images():
     Returns:
         dict: A dictionary of shortname to image paths.
     """
+    if StateTracker.get_model().requires_validation_edit_captions():
+        return retrieve_validation_edit_images()
+
     args = StateTracker.get_args()
+    requires_cond_input = any(
+        [
+            StateTracker.get_model().requires_conditioning_validation_inputs(),
+            args.controlnet,
+            args.control,
+        ]
+    )
     data_backends = StateTracker.get_data_backends(
-        _type="conditioning" if args.controlnet else "image"
+        _type=(
+            StateTracker.get_model().conditioning_validation_dataset_type()
+            if requires_cond_input
+            else "image"
+        )
     )
     validation_data_backend_id = args.eval_dataset_id
     validation_set = []
@@ -186,7 +196,9 @@ def retrieve_validation_images():
         should_skip_dataset = data_backend_config.get("disable_validation", False)
         logger.debug(f"Backend {_data_backend}: {data_backend}")
         if "id" not in data_backend or (
-            args.controlnet and data_backend.get("dataset_type", None) != "conditioning"
+            requires_cond_input
+            and data_backend.get("dataset_type", None)
+            != StateTracker.get_model().conditioning_validation_dataset_type()
         ):
             logger.debug(
                 f"Skipping data backend: {_data_backend} dataset_type {data_backend.get('dataset_type', None)}"
@@ -203,9 +215,16 @@ def retrieve_validation_images():
             validation_samples_from_sampler = data_backend[
                 "sampler"
             ].retrieve_validation_set(batch_size=args.num_eval_images)
-            if "stage2" in args.model_type:
+            validation_input_image_pixel_edge_len = (
+                StateTracker.get_model().validation_image_input_edge_length()
+            )
+            if validation_input_image_pixel_edge_len is not None:
+                logger.debug(
+                    f"Resized validation image so that pixel edge length is {validation_input_image_pixel_edge_len}."
+                )
                 validation_samples_from_sampler = resize_validation_images(
-                    validation_samples_from_sampler, edge_length=64
+                    validation_samples_from_sampler,
+                    edge_length=validation_input_image_pixel_edge_len,
                 )
 
             validation_set.extend(validation_samples_from_sampler)
@@ -213,12 +232,96 @@ def retrieve_validation_images():
             logger.warning(
                 f"Data backend {data_backend['id']} does not have a sampler. Skipping."
             )
+    logger.info(f"Collected {len(validation_set)} validation image inputs.")
     return validation_set
 
 
-def prepare_validation_prompt_list(args, embed_cache):
-    validation_negative_prompt_embeds = None
-    validation_negative_pooled_embeds = None
+def retrieve_validation_edit_images() -> list[tuple[str, str, Image.Image]]:
+    """
+    Returns [(shortname, *edited-scene caption*, reference_image), ...]
+    for models that need **edit** validation.
+
+    Logic
+    -----
+    • loop over *image* datasets that have a sampler
+    • for every deterministic validation sample returned by the sampler
+      – grab its original file path from metadata
+      – ask the dataset’s *registered* conditioning backend for the
+        counterpart via `get_conditioning_sample()`
+      – add the trio to output
+    """
+    model = StateTracker.get_model()
+    if not model.requires_validation_edit_captions():
+        return []  # no-op for ordinary models
+
+    args = StateTracker.get_args()
+    validation_set = []
+
+    # ---------- iterate over IMAGE datasets ---------------------------------
+    for backend_id, backend in StateTracker.get_data_backends(_type="image").items():
+        sampler = backend.get("sampler")
+        if sampler is None:  # nothing to iterate over
+            continue
+
+        # each backend should know which conditioning dataset is linked to it
+        cond_backend = backend.get("conditioning_data")
+        cond_backend_id = cond_backend.get("id") if cond_backend else None
+        if cond_backend_id is None:
+            logger.debug("No conditioning backend configured for this image dataset.")
+            continue
+
+        logger.debug(f"Backend id: {backend_id}, cond_backend_id: {cond_backend_id}")
+        cond_backend = StateTracker.get_data_backend(cond_backend_id)
+        if cond_backend is None:
+            logger.debug(f"Conditioning backend {cond_backend_id} not found.")
+            continue  # mis-configuration → skip
+
+        # deterministic slice for validation
+        for sample in sampler.retrieve_validation_set(batch_size=args.num_eval_images):
+            # sample == (shortname, edited_prompt, pil_image)
+            shortname, edited_prompt, sample_path, _ = sample
+
+            # original relative file-path comes from metadata backend
+            try:
+                meta = StateTracker.get_metadata_by_filepath(
+                    sample_path,
+                    data_backend_id=backend_id,
+                    search_dataset_types=["conditioning"],
+                )
+                image_dataset_dir_prefix = sampler.metadata_backend.instance_data_dir
+                if (
+                    image_dataset_dir_prefix is not None
+                    and image_dataset_dir_prefix in sample_path
+                ):
+                    rel_path = sample_path.replace(image_dataset_dir_prefix, "")
+                    # remove trailing '/'
+                    rel_path = rel_path.lstrip("/")
+                    logger.debug(f"Removed prefix, got relative path: {rel_path}")
+                logger.debug(f"Metadata: {meta}")
+            except Exception as e:
+                continue  # metadata missing → skip
+
+            conditioning_dir_prefix = cond_backend[
+                "sampler"
+            ].metadata_backend.instance_data_dir
+            cond_sample_path = os.path.join(conditioning_dir_prefix, rel_path)
+            cond_sample = cond_backend["sampler"].get_conditioning_sample(rel_path)
+
+            if cond_sample is None:
+                logger.warning(
+                    f"Conditioning sample for {rel_path} not found in {cond_backend_id}."
+                )
+                continue
+
+            reference_img = cond_sample.image
+            logger.debug(f"Validation sample path: {cond_sample_path}")
+            validation_set.append((shortname, edited_prompt, reference_img))
+
+    logger.info(f"Collected {len(validation_set)} edit-validation samples.")
+    return validation_set
+
+
+def prepare_validation_prompt_list(args, embed_cache, model):
     validation_prompts = (
         [""] if not StateTracker.get_args().validation_disable_unconditional else []
     )
@@ -234,8 +337,10 @@ def prepare_validation_prompt_list(args, embed_cache):
     model_type = embed_cache.model_type
     validation_sample_images = None
     if (
-        "deepfloyd-stage2" in args.model_type
+        ("deepfloyd" in args.model_family and str(args.model_flavour).startswith("ii-"))
+        or model.requires_conditioning_validation_inputs()
         or args.controlnet
+        or args.control
         or args.validation_using_datasets
     ):
         # Now, we prepare the DeepFloyd upscaler image inputs so that we can calculate their prompts.
@@ -246,10 +351,19 @@ def prepare_validation_prompt_list(args, embed_cache):
             # Collect the prompts for the validation images.
             for _validation_sample in tqdm(
                 validation_sample_images,
-                ncols=100,
+                ncols=125,
                 desc="Precomputing validation image embeds",
             ):
-                _, validation_prompt, _ = _validation_sample
+                if (
+                    isinstance(_validation_sample, tuple)
+                    and len(_validation_sample) == 3
+                ):
+                    _, validation_prompt, _ = _validation_sample
+                elif (
+                    isinstance(_validation_sample, tuple)
+                    and len(_validation_sample) == 4
+                ):
+                    _, validation_prompt, _, _ = _validation_sample
                 embed_cache.compute_embeddings_for_prompts(
                     [validation_prompt], load_from_cache=False
                 )
@@ -263,7 +377,7 @@ def prepare_validation_prompt_list(args, embed_cache):
         for shortname, prompt in tqdm(
             prompt_library.items(),
             leave=False,
-            ncols=100,
+            ncols=125,
             desc="Precomputing validation prompt embeddings",
         ):
             embed_cache.compute_embeddings_for_prompts(
@@ -271,12 +385,13 @@ def prepare_validation_prompt_list(args, embed_cache):
             )
             validation_prompts.append(prompt)
             validation_shortnames.append(shortname)
+
     if args.user_prompt_library is not None:
         user_prompt_library = PromptHandler.load_user_prompts(args.user_prompt_library)
         for shortname, prompt in tqdm(
             user_prompt_library.items(),
             leave=False,
-            ncols=100,
+            ncols=125,
             desc="Precomputing user prompt library embeddings",
         ):
             # move_text_encoders(embed_cache.text_encoders, embed_cache.accelerator.device)
@@ -294,118 +409,24 @@ def prepare_validation_prompt_list(args, embed_cache):
         embed_cache.compute_embeddings_for_prompts(
             [args.validation_prompt], is_validation=True, load_from_cache=False
         )
-
-    # Compute negative embed for validation prompts, if any are set.
+    # Compute negative embed for validation prompts, if any are set, so that it's stored before we unload the text encoder.
     if validation_prompts:
         logger.info("Precomputing the negative prompt embed for validations.")
-        if model_type == "sdxl" or model_type == "sd3" or model_type == "kolors":
-            (
-                validation_negative_prompt_embeds,
-                validation_negative_pooled_embeds,
-            ) = embed_cache.compute_embeddings_for_prompts(
+        model.log_model_devices()
+        validation_negative_prompt_text_encoder_output = (
+            embed_cache.compute_embeddings_for_prompts(
                 [StateTracker.get_args().validation_negative_prompt],
                 is_validation=True,
                 load_from_cache=False,
             )
-            return (
-                validation_prompts,
-                validation_shortnames,
-                validation_negative_prompt_embeds,
-                validation_negative_pooled_embeds,
-            )
-        elif model_type == "legacy":
-            validation_negative_prompt_embeds = (
-                embed_cache.compute_embeddings_for_prompts(
-                    [StateTracker.get_args().validation_negative_prompt],
-                    load_from_cache=False,
-                )
-            )
+        )
 
-            return (
-                validation_prompts,
-                validation_shortnames,
-                validation_negative_prompt_embeds,
-                None,
-            )
-        elif model_type in ["pixart_sigma", "smoldit", "sana", "ltxvideo", "wan"]:
-            # we use the legacy encoder but we return no pooled embeds.
-            validation_negative_prompt_embeds = embed_cache.compute_embeddings_for_prompts(
-                [StateTracker.get_args().validation_negative_prompt],
-                load_from_cache=False,
-                is_negative_prompt=True,  # sana needs this to disable Complex Human Instruction on negative embed generation
-            )
-
-            return (
-                validation_prompts,
-                validation_shortnames,
-                validation_negative_prompt_embeds,
-                None,
-            )
-        elif model_type == "flux":
-            (
-                validation_negative_prompt_embeds,
-                validation_negative_pooled_embeds,
-                validation_negative_time_ids,
-                _,
-            ) = embed_cache.compute_embeddings_for_prompts(
-                [StateTracker.get_args().validation_negative_prompt],
-                load_from_cache=False,
-            )
-            return (
-                validation_prompts,
-                validation_shortnames,
-                validation_negative_prompt_embeds,
-                validation_negative_pooled_embeds,
-                validation_negative_time_ids,
-            )
-        else:
-            raise ValueError(f"Unknown model type '{model_type}'")
-
-
-def parse_validation_resolution(input_str: str) -> tuple:
-    """
-    If the args.validation_resolution:
-     - is an int, we'll treat it as height and width square aspect
-     - if it has an x in it, we will split and treat as WIDTHxHEIGHT
-     - if it has comma, we will split and treat each value as above
-    """
-    if isinstance(input_str, int) or input_str.isdigit():
-        if (
-            "deepfloyd-stage2" in StateTracker.get_args().model_type
-            and int(input_str) < 256
-        ):
-            raise ValueError(
-                "Cannot use less than 256 resolution for DeepFloyd stage 2."
-            )
-        return (input_str, input_str)
-    if "x" in input_str:
-        pieces = input_str.split("x")
-        if "deepfloyd-stage2" in StateTracker.get_args().model_type and (
-            int(pieces[0]) < 256 or int(pieces[1]) < 256
-        ):
-            raise ValueError(
-                "Cannot use less than 256 resolution for DeepFloyd stage 2."
-            )
-        return (int(pieces[0]), int(pieces[1]))
-
-
-def get_validation_resolutions():
-    """
-    If the args.validation_resolution:
-     - is an int, we'll treat it as height and width square aspect
-     - if it has an x in it, we will split and treat as WIDTHxHEIGHT
-     - if it has comma, we will split and treat each value as above
-    """
-    validation_resolution_parameter = StateTracker.get_args().validation_resolution
-    if (
-        type(validation_resolution_parameter) is str
-        and "," in validation_resolution_parameter
-    ):
-        return [
-            parse_validation_resolution(res)
-            for res in validation_resolution_parameter.split(",")
-        ]
-    return [parse_validation_resolution(validation_resolution_parameter)]
+    logger.info("Completed validation prompt gathering.")
+    return {
+        "validation_prompts": validation_prompts,
+        "validation_shortnames": validation_shortnames,
+        "validation_sample_images": validation_sample_images,
+    }
 
 
 def get_validation_resolutions():
@@ -435,7 +456,7 @@ def parse_validation_resolution(input_str: str) -> tuple:
      - if it has comma, we will split and treat each value as above
     """
     is_df_ii = (
-        True if "deepfloyd-stage2" in StateTracker.get_args().model_type else False
+        True if str(StateTracker.get_args().model_flavour).startswith("ii-") else False
     )
     if isinstance(input_str, int) or input_str.isdigit():
         if is_df_ii and int(input_str) < 256:
@@ -456,25 +477,14 @@ class Validation:
     def __init__(
         self,
         accelerator,
-        unet,
-        transformer,
+        model: ModelFoundation,
+        distiller,
         args,
-        validation_prompts,
-        validation_shortnames,
-        text_encoder_1,
-        tokenizer,
+        validation_prompt_metadata,
         vae_path,
         weight_dtype,
         embed_cache,
-        validation_negative_pooled_embeds,
-        validation_negative_prompt_embeds,
-        text_encoder_2,
-        tokenizer_2,
         ema_model,
-        vae,
-        controlnet=None,
-        text_encoder_3=None,
-        tokenizer_3=None,
         is_deepspeed: bool = False,
         model_evaluator=None,
         trainable_parameters=None,
@@ -482,53 +492,40 @@ class Validation:
         self.trainable_parameters = trainable_parameters
         self.accelerator = accelerator
         self.prompt_handler = None
-        self.unet = unet
-        self.transformer = transformer
-        self.controlnet = controlnet
-        self.args = args
+        self.unet, self.transformer = None, None
+        self.model = model
+        self.distiller = distiller
+        if args.controlnet:
+            self.controlnet = model.get_trained_component()
+        elif "unet" in str(self.model.get_trained_component().__class__).lower():
+            self.unet = self.model.get_trained_component()
+        elif "transformer" in str(self.model.get_trained_component().__class__).lower():
+            self.transformer = self.model.get_trained_component()
+        self.config = args
         self.save_dir = os.path.join(args.output_dir, "validation_images")
         if not os.path.exists(self.save_dir):
             os.makedirs(self.save_dir, exist_ok=True)
         self.global_step = None
         self.global_resume_step = None
-        self.text_encoder_1 = text_encoder_1
-        self.tokenizer_1 = tokenizer
-        self.text_encoder_2 = text_encoder_2
-        self.tokenizer_2 = tokenizer_2
-        self.vae_path = vae_path
-        self.validation_prompts = validation_prompts
-        self.validation_shortnames = validation_shortnames
+        self.validation_prompt_metadata = validation_prompt_metadata
         self.validation_images = None
         self.weight_dtype = weight_dtype
         self.embed_cache = embed_cache
-        self.validation_negative_prompt_mask = None
-        self.validation_negative_pooled_embeds = validation_negative_pooled_embeds
-        self.validation_negative_prompt_embeds = (
-            validation_negative_prompt_embeds
-            if (
-                type(validation_negative_prompt_embeds) is not list
-                and type(validation_negative_prompt_embeds) is not tuple
-            )
-            else validation_negative_prompt_embeds[0]
-        )
         self.ema_model = ema_model
         self.ema_enabled = False
-        self.vae = vae
-        self.pipeline = None
-        self.deepfloyd = True if "deepfloyd" in self.args.model_type else False
+        self.deepfloyd = True if "deepfloyd" in self.config.model_family else False
         self.deepfloyd_stage2 = (
-            True if "deepfloyd-stage2" in self.args.model_type else False
+            True if str(self.config.model_flavour).startswith("ii-") else False
         )
         self._discover_validation_input_samples()
         self.validation_resolutions = (
-            get_validation_resolutions() if not self.deepfloyd_stage2 else ["base-256"]
+            get_validation_resolutions() if not self.deepfloyd_stage2 else [(256, 256)]
         )
-        self.text_encoder_3 = text_encoder_3
-        self.tokenizer_3 = tokenizer_3
         self.flow_matching = (
-            self.args.model_family == "sd3"
-            and self.args.flow_matching_loss != "diffusion"
-        ) or self.args.model_family == "flux"
+            True
+            if self.model.PREDICTION_TYPE is PredictionTypes.FLOW_MATCHING
+            else False
+        )
         self.deepspeed = is_deepspeed
         if is_deepspeed:
             if args.use_ema:
@@ -545,14 +542,14 @@ class Validation:
         )
         self.model_evaluator = model_evaluator
         if self.model_evaluator is not None:
-            logger.info(f"Using model evaluator: {self.model_evaluator}")
+            logger.debug(f"Using model evaluator: {self.model_evaluator}")
         self._update_state()
         self.eval_scores = {}
 
     def _validation_seed_source(self):
-        if self.args.validation_seed_source == "gpu":
+        if self.config.validation_seed_source == "gpu":
             return self.inference_device
-        elif self.args.validation_seed_source == "cpu":
+        elif self.config.validation_seed_source == "cpu":
             return "cpu"
         else:
             raise Exception("Unknown validation seed source. Options: cpu, gpu")
@@ -560,7 +557,7 @@ class Validation:
     def _get_generator(self):
         _validation_seed_source = self._validation_seed_source()
         _generator = torch.Generator(device=_validation_seed_source).manual_seed(
-            self.args.validation_seed or self.args.seed or 0
+            self.config.validation_seed or self.config.seed or 0
         )
         return _generator
 
@@ -571,48 +568,7 @@ class Validation:
         Returns:
             None
         """
-        self.text_encoder_1 = None
-        self.text_encoder_2 = None
-        self.text_encoder_3 = None
-
-    def init_vae(self):
-
-        args = StateTracker.get_args()
-        vae_path = (
-            args.pretrained_model_name_or_path
-            if args.pretrained_vae_model_name_or_path is None
-            else args.pretrained_vae_model_name_or_path
-        )
-        precached_vae = StateTracker.get_vae()
-        logger.debug(
-            f"Was the VAE loaded? {precached_vae if precached_vae is None else 'Yes'}"
-        )
-        if self.args.model_family == "sana":
-            from diffusers import AutoencoderDC as AutoencoderClass
-        elif self.args.model_family == "wan":
-            from diffusers import AutoencoderKLWan as AutoencoderClass
-        elif self.args.model_family == "ltxvideo":
-            from diffusers import AutoencoderKLLTXVideo as AutoencoderClass
-        else:
-            from diffusers import AutoencoderKL as AutoencoderClass
-        self.vae = precached_vae
-        if self.vae is None:
-            logger.info(f"Initialising {AutoencoderClass}")
-            with ContextManagers(deepspeed_zero_init_disabled_context_manager()):
-                self.vae = AutoencoderClass.from_pretrained(
-                    vae_path,
-                    subfolder=(
-                        "vae"
-                        if args.pretrained_vae_model_name_or_path is None
-                        else None
-                    ),
-                    revision=args.revision,
-                    force_upcast=False,
-                ).to(self.inference_device)
-        StateTracker.set_vae(self.vae)
-
-        # logger.info(f"VAE type: {type(self.vae)}")
-        return self.vae
+        self.model.unload_text_encoder()
 
     def _discover_validation_input_samples(self):
         """
@@ -624,8 +580,10 @@ class Validation:
         self.validation_image_inputs = None
         if (
             self.deepfloyd_stage2
-            or self.args.validation_using_datasets
-            or self.args.controlnet
+            or self.config.validation_using_datasets
+            or self.config.controlnet
+            or self.config.control
+            or self.model.requires_conditioning_validation_inputs()
         ):
             self.validation_image_inputs = retrieve_validation_images()
             # Validation inputs are in the format of a list of tuples:
@@ -635,230 +593,52 @@ class Validation:
             )
 
     def _pipeline_cls(self):
-        model_type = StateTracker.get_model_family()
-        if model_type == "sdxl":
-            if self.args.controlnet:
-                from diffusers.pipelines import StableDiffusionXLControlNetPipeline
-
-                return StableDiffusionXLControlNetPipeline
-            if self.args.validation_using_datasets:
-                return StableDiffusionXLImg2ImgPipeline
-            return StableDiffusionXLPipeline
-        elif model_type == "flux":
-            from helpers.models.flux import FluxPipeline
-
-            if self.args.controlnet:
-                raise NotImplementedError("Flux ControlNet is not yet supported.")
-            if self.args.validation_using_datasets:
-                raise NotImplementedError(
-                    "Flux inference validation using img2img is not yet supported. Please remove --validation_using_datasets."
-                )
-            return FluxPipeline
-        elif model_type == "kolors":
-            if self.args.controlnet:
-                raise NotImplementedError("Kolors ControlNet is not yet supported.")
-            if self.args.validation_using_datasets:
-                try:
-                    from helpers.kolors.pipeline import KolorsImg2ImgPipeline
-                except:
-                    logger.error(
-                        "Kolors pipeline requires the latest version of Diffusers."
+        if self.model is not None:
+            if self.config.validation_using_datasets:
+                if PipelineTypes.IMG2IMG not in self.model.PIPELINE_CLASSES:
+                    raise ValueError(
+                        f"Cannot run {self.model.MODEL_CLASS} in Img2Img mode for validation."
                     )
-                return KolorsImg2ImgPipeline
-            try:
-                from helpers.kolors.pipeline import KolorsPipeline
-            except Exception:
-                logger.error(
-                    "Kolors pipeline requires the latest version of Diffusers."
-                )
-            return KolorsPipeline
-        elif model_type == "legacy":
-            if self.deepfloyd_stage2:
-                from diffusers.pipelines import IFSuperResolutionPipeline
+            if self.config.controlnet:
+                if PipelineTypes.CONTROLNET not in self.model.PIPELINE_CLASSES:
+                    raise ValueError(
+                        f"Cannot run {self.model.MODEL_CLASS} in ControlNet mode for validation."
+                    )
+            if self.config.control:
+                if PipelineTypes.CONTROL not in self.model.PIPELINE_CLASSES:
+                    raise ValueError(
+                        f"Cannot run {self.model.MODEL_CLASS} in Control mode for validation."
+                    )
 
-                return IFSuperResolutionPipeline
-            return StableDiffusionPipeline
-        elif model_type == "sd3":
-            if self.args.controlnet:
-                raise Exception("SD3 ControlNet is not yet supported.")
-            if self.args.validation_using_datasets:
-                return StableDiffusion3Img2ImgPipeline
-            return StableDiffusion3Pipeline
-        elif model_type == "pixart_sigma":
-            if self.args.controlnet:
-                raise Exception(
-                    "PixArt Sigma ControlNet inference validation is not yet supported."
-                )
-            if self.args.validation_using_datasets:
-                raise Exception(
-                    "PixArt Sigma inference validation using img2img is not yet supported. Please remove --validation_using_datasets."
-                )
-            from helpers.models.pixart.pipeline import PixArtSigmaPipeline
-
-            return PixArtSigmaPipeline
-        elif model_type == "smoldit":
-            from helpers.models.smoldit import SmolDiTPipeline
-
-            return SmolDiTPipeline
-        elif model_type == "sana":
-            from diffusers import SanaPipeline
-
-            return SanaPipeline
-        elif model_type == "ltxvideo":
-            from diffusers import LTXPipeline
-
-            return LTXPipeline
-        elif model_type == "wan":
-            from helpers.models.wan.pipeline import WanPipeline
-
-            return WanPipeline
-        else:
-            raise NotImplementedError(
-                f"Model type {model_type} not implemented for validation."
-            )
+        if self.config.validation_using_datasets:
+            return self.model.PIPELINE_CLASSES[PipelineTypes.IMG2IMG]
+        if self.config.controlnet:
+            return self.model.PIPELINE_CLASSES[PipelineTypes.CONTROLNET]
+        if self.config.control:
+            return self.model.PIPELINE_CLASSES[PipelineTypes.CONTROL]
+        return self.model.PIPELINE_CLASSES[PipelineTypes.TEXT2IMG]
 
     def _gather_prompt_embeds(self, validation_prompt: str):
-        prompt_embeds = {}
-        current_validation_prompt_mask = None
-        if (
-            StateTracker.get_model_family() == "sdxl"
-            or StateTracker.get_model_family() == "sd3"
-            or StateTracker.get_model_family() == "kolors"
-            or StateTracker.get_model_family() == "flux"
-        ):
-            _embed = self.embed_cache.compute_embeddings_for_prompts(
-                [validation_prompt]
-            )
-            current_validation_time_ids = None
-            if len(_embed) == 2:
-                (
-                    current_validation_prompt_embeds,
-                    current_validation_pooled_embeds,
-                ) = _embed
-            elif len(_embed) == 3:
-                (
-                    current_validation_prompt_embeds,
-                    current_validation_pooled_embeds,
-                    current_validation_time_ids,
-                ) = _embed
-            elif len(_embed) == 4:
-                (
-                    current_validation_prompt_embeds,
-                    current_validation_pooled_embeds,
-                    current_validation_time_ids,
-                    current_validation_prompt_mask,
-                ) = _embed
-            else:
-                raise ValueError(
-                    f"Unexpected number of embeddings returned from cache: {_embed}"
-                )
-            current_validation_pooled_embeds = current_validation_pooled_embeds.to(
-                device=self.inference_device, dtype=self.weight_dtype
-            )
-            if current_validation_time_ids is not None:
-                current_validation_time_ids = current_validation_time_ids.to(
-                    device=self.inference_device, dtype=self.weight_dtype
-                )
-            self.validation_negative_pooled_embeds = (
-                self.validation_negative_pooled_embeds.to(
-                    device=self.inference_device, dtype=self.weight_dtype
-                )
-            )
-            prompt_embeds["pooled_prompt_embeds"] = current_validation_pooled_embeds.to(
-                device=self.inference_device, dtype=self.weight_dtype
-            )
-            prompt_embeds["negative_pooled_prompt_embeds"] = (
-                self.validation_negative_pooled_embeds
-            )
-            # if current_validation_time_ids is not None:
-            #     prompt_embeds["time_ids"] = current_validation_time_ids
-        elif StateTracker.get_model_family() in [
-            "legacy",
-            "pixart_sigma",
-            "smoldit",
-            "sana",
-            "ltxvideo",
-            "wan",
-        ]:
-            self.validation_negative_pooled_embeds = None
-            current_validation_pooled_embeds = None
-            current_validation_prompt_embeds = (
-                self.embed_cache.compute_embeddings_for_prompts([validation_prompt])
-            )
-            if StateTracker.get_model_family() in [
-                "pixart_sigma",
-                "smoldit",
-                "sana",
-                "ltxvideo",
-                "wan",
-            ]:
-                current_validation_prompt_embeds, current_validation_prompt_mask = (
-                    current_validation_prompt_embeds
-                )
-                current_validation_prompt_embeds = current_validation_prompt_embeds[
-                    0
-                ].to(device=self.inference_device, dtype=self.weight_dtype)
-                if (
-                    type(self.validation_negative_prompt_embeds) is tuple
-                    or type(self.validation_negative_prompt_embeds) is list
-                ):
-                    (
-                        self.validation_negative_prompt_embeds,
-                        self.validation_negative_prompt_mask,
-                    ) = self.validation_negative_prompt_embeds[0]
-            else:
-                current_validation_prompt_embeds = current_validation_prompt_embeds[
-                    0
-                ].to(device=self.inference_device, dtype=self.weight_dtype)
-            # logger.debug(
-            #     f"Validations received the prompt embed: ({type(current_validation_prompt_embeds)}) positive={current_validation_prompt_embeds.shape if type(current_validation_prompt_embeds) is not list else current_validation_prompt_embeds[0].shape},"
-            #     f" ({type(self.validation_negative_prompt_embeds)}) negative={self.validation_negative_prompt_embeds.shape if type(self.validation_negative_prompt_embeds) is not list else self.validation_negative_prompt_embeds[0].shape}"
-            # )
-            # logger.debug(
-            #     f"Dtypes: {current_validation_prompt_embeds.dtype}, {self.validation_negative_prompt_embeds.dtype}"
-            # )
-        else:
-            raise NotImplementedError(
-                f"Model type {StateTracker.get_model_family()} not implemented for validation."
-            )
-
-        current_validation_prompt_embeds = current_validation_prompt_embeds.to(
-            device=self.inference_device, dtype=self.weight_dtype
+        prompt_embed = self.embed_cache.compute_embeddings_for_prompts(
+            [validation_prompt]
         )
-        self.validation_negative_prompt_embeds = (
-            self.validation_negative_prompt_embeds.to(
-                device=self.inference_device, dtype=self.weight_dtype
-            )
-        )
-        # when sampling unconditional guidance, you should only zero one or the other prompt, and not both.
-        # we'll assume that the user has a negative prompt, so that the unconditional sampling works.
-        # the positive prompt embed is zeroed out for SDXL at the time of it being placed into the cache.
-        # the embeds are not zeroed out for any other model, including Stable Diffusion 3.
-        prompt_embeds["prompt_embeds"] = current_validation_prompt_embeds
-        prompt_embeds["negative_prompt_embeds"] = self.validation_negative_prompt_embeds
-        if StateTracker.get_model_family() in [
-            "pixart_sigma",
-            "smoldit",
-            "sana",
-            "ltxvideo",
-        ] or (
-            StateTracker.get_model_family() == "flux"
-            and StateTracker.get_args().flux_attention_masked_training
-        ):
-            logger.debug(
-                f"mask: {current_validation_prompt_mask.shape if type(current_validation_prompt_mask) is torch.Tensor else None}"
-            )
-            assert current_validation_prompt_mask is not None
-            prompt_embeds["prompt_mask"] = current_validation_prompt_mask
-            prompt_embeds["negative_mask"] = self.validation_negative_prompt_mask
+        if prompt_embed is None:
+            return
 
-        return prompt_embeds
+        prompt_embed = {
+            k: v.to(self.inference_device) if hasattr(v, "to") else v
+            for k, v in prompt_embed.items()
+        }
+
+        return self.model.convert_text_embed_for_pipeline(prompt_embed)
 
     def _benchmark_path(self, benchmark: str = "base_model"):
         # does the benchmark directory exist?
-        if not os.path.exists(os.path.join(self.args.output_dir, "benchmarks")):
-            os.makedirs(os.path.join(self.args.output_dir, "benchmarks"), exist_ok=True)
-        return os.path.join(self.args.output_dir, "benchmarks", benchmark)
+        if not os.path.exists(os.path.join(self.config.output_dir, "benchmarks")):
+            os.makedirs(
+                os.path.join(self.config.output_dir, "benchmarks"), exist_ok=True
+            )
+        return os.path.join(self.config.output_dir, "benchmarks", benchmark)
 
     def stitch_benchmark_image(
         self,
@@ -995,17 +775,13 @@ class Validation:
                         )
                     )
                 elif type(image) is list:
-                    # maybe video
-                    if self.args.model_family in ["ltxvideo", "wan"]:
-                        from diffusers.utils.export_utils import export_to_video
+                    from diffusers.utils.export_utils import export_to_video
 
-                        export_to_video(
-                            image,
-                            os.path.join(
-                                base_model_benchmark, f"{shortname}_{idx}.mp4"
-                            ),
-                            fps=self.args.framerate,
-                        )
+                    export_to_video(
+                        image,
+                        os.path.join(base_model_benchmark, f"{shortname}_{idx}.mp4"),
+                        fps=self.config.framerate,
+                    )
 
     def _update_state(self):
         """Updates internal state with the latest from StateTracker."""
@@ -1021,7 +797,7 @@ class Validation:
         # a wrapper for should_perform_intermediary_validation that can run in the training loop
         self._update_state()
         return self.should_perform_intermediary_validation(
-            step, self.validation_prompts, validation_type
+            step, self.validation_prompt_metadata, validation_type
         ) or (step == 0 and validation_type == "base_model")
 
     def run_validations(
@@ -1033,7 +809,7 @@ class Validation:
     ):
         self._update_state()
         would_do_intermediary_validation = self.should_perform_intermediary_validation(
-            step, self.validation_prompts, validation_type
+            step, self.validation_prompt_metadata, validation_type
         ) or (step == 0 and validation_type == "base_model")
         logger.debug(
             f"Should evaluate: {would_do_intermediary_validation}, force evaluation: {force_evaluation}, skip execution: {skip_execution}"
@@ -1057,7 +833,7 @@ class Validation:
             logger.debug("Starting validation process...")
             diffusers.utils.logging._tqdm_active = False
             self.setup_pipeline(validation_type)
-            if self.pipeline is None:
+            if self.model.pipeline is None:
                 logger.error(
                     "Not able to run validations, we did not obtain a valid pipeline."
                 )
@@ -1078,8 +854,8 @@ class Validation:
     ):
         should_do_intermediary_validation = (
             validation_prompts
-            and self.global_step % self.args.validation_steps == 0
-            and step % self.args.gradient_accumulation_steps == 0
+            and self.global_step % self.config.validation_steps == 0
+            and step % self.config.gradient_accumulation_steps == 0
             and self.global_step > self.global_resume_step
         )
         return should_do_intermediary_validation and (
@@ -1087,243 +863,123 @@ class Validation:
         )
 
     def setup_scheduler(self):
+        if self.distiller is not None:
+            distillation_scheduler = self.distiller.get_scheduler()
+            if distillation_scheduler is not None:
+                self.model.pipeline.scheduler = distillation_scheduler
+                return distillation_scheduler
         scheduler_args = {
-            "prediction_type": self.args.prediction_type,
+            "prediction_type": self.config.prediction_type,
         }
-        if self.flow_matching and not self.args.model_family not in ["wan", "sana"]:
-            # NO TOUCHIE FOR FLOW-MATCHING.
-            # Touchie for sana though. It needs a new scheduler on every inference.
-            return
-        elif self.args.model_family == "sana":
-            self.args.validation_noise_scheduler = "sana"
-        elif self.args.model_family in ["ltxvideo", "wan"]:
-            if self.args.validation_noise_scheduler is None:
-                # Diffusers repo uses UniPC by default.
-                if self.args.model_family == "ltxvideo":
-                    self.args.validation_noise_scheduler = "flow-match"
-                else:
-                    self.args.validation_noise_scheduler = "unipc"
-            if self.args.validation_noise_scheduler == "flow-match":
+        if self.config.model_family == "sana":
+            self.config.validation_noise_scheduler = "sana"
+        elif (
+            getattr(self.model, "DEFAULT_NOISE_SCHEDULER", None) is not None
+            and self.config.validation_noise_scheduler is None
+        ):
+            # set the default
+            self.config.validation_noise_scheduler = self.model.DEFAULT_NOISE_SCHEDULER
+        if self.model.PREDICTION_TYPE.value == "flow_matching":
+            # some flow-matching adjustments should be made for euler and unipc video model generations.
+            if self.config.validation_noise_scheduler in ["flow_matching", "euler"]:
                 # The Beta schedule looks WAY better...
                 scheduler_args["use_beta_sigmas"] = True
-                scheduler_args["shift"] = self.args.flow_schedule_shift
-            if self.args.validation_noise_scheduler == "unipc":
+                scheduler_args["shift"] = self.config.flow_schedule_shift
+            if self.config.validation_noise_scheduler in ["flow_unipc", "unipc"]:
                 scheduler_args["prediction_type"] = "flow_prediction"
                 scheduler_args["use_flow_sigmas"] = True
                 scheduler_args["num_train_timesteps"] = 1000
-                scheduler_args["flow_shift"] = self.args.flow_schedule_shift
+                scheduler_args["flow_shift"] = self.config.flow_schedule_shift
 
-        if self.args.validation_noise_scheduler is None:
+        if self.config.validation_noise_scheduler is None:
+            # if the user or model config has not supplied one, we just allow pipeline to do defaults.
             return
 
+        if self.config.prediction_type is not None:
+            scheduler_args["prediction_type"] = self.config.prediction_type
+
         if (
-            self.pipeline is not None
-            and "variance_type" in self.pipeline.scheduler.config
+            self.model.pipeline is not None
+            and "variance_type" in self.model.pipeline.scheduler.config
         ):
-            variance_type = self.pipeline.scheduler.config.variance_type
+            variance_type = self.model.pipeline.scheduler.config.variance_type
 
             if variance_type in ["learned", "learned_range"]:
                 variance_type = "fixed_small"
 
             scheduler_args["variance_type"] = variance_type
-        if self.deepfloyd:
-            self.args.validation_noise_scheduler = "ddpm"
+
         scheduler = SCHEDULER_NAME_MAP[
-            self.args.validation_noise_scheduler
+            self.config.validation_noise_scheduler
         ].from_pretrained(
-            self.args.pretrained_model_name_or_path,
+            self.config.pretrained_model_name_or_path,
             subfolder="scheduler",
-            revision=self.args.revision,
-            timestep_spacing=self.args.inference_scheduler_timestep_spacing,
-            rescale_betas_zero_snr=self.args.rescale_betas_zero_snr,
+            revision=self.config.revision,
+            timestep_spacing=self.config.inference_scheduler_timestep_spacing,
+            rescale_betas_zero_snr=self.config.rescale_betas_zero_snr,
             **scheduler_args,
         )
-        if self.pipeline is not None:
-            self.pipeline.scheduler = scheduler
+        if self.model.pipeline is not None:
+            self.model.pipeline.scheduler = scheduler
         return scheduler
 
     def setup_pipeline(self, validation_type):
         if hasattr(self.accelerator, "_lycoris_wrapped_network"):
             self.accelerator._lycoris_wrapped_network.set_multiplier(
-                float(getattr(self.args, "validation_lycoris_strength", 1.0))
+                float(getattr(self.config, "validation_lycoris_strength", 1.0))
             )
 
-        if self.pipeline is None:
-            pipeline_cls = self._pipeline_cls()
-            extra_pipeline_kwargs = {
-                "text_encoder": self.text_encoder_1,
-                "tokenizer": self.tokenizer_1,
-                "vae": self.vae,
-            }
-            if self.args.model_family in ["legacy"]:
-                extra_pipeline_kwargs["safety_checker"] = None
-            if self.args.model_family in ["sd3", "sdxl", "flux"]:
-                extra_pipeline_kwargs["text_encoder_2"] = None
-            if self.args.model_family in ["sd3"]:
-                extra_pipeline_kwargs["text_encoder_3"] = None
-            if type(pipeline_cls) is StableDiffusionXLPipeline:
-                del extra_pipeline_kwargs["text_encoder"]
-                del extra_pipeline_kwargs["tokenizer"]
-                if validation_type == "final":
-                    if self.text_encoder_1 is not None:
-                        extra_pipeline_kwargs["text_encoder_1"] = unwrap_model(
-                            self.accelerator, self.text_encoder_1
-                        )
-                        extra_pipeline_kwargs["tokenizer_1"] = self.tokenizer_1
-                        if self.text_encoder_2 is not None:
-                            extra_pipeline_kwargs["text_encoder_2"] = unwrap_model(
-                                self.accelerator, self.text_encoder_2
-                            )
-                            extra_pipeline_kwargs["tokenizer_2"] = self.tokenizer_2
-                else:
-                    extra_pipeline_kwargs["text_encoder_1"] = None
-                    extra_pipeline_kwargs["tokenizer_1"] = None
-                    extra_pipeline_kwargs["text_encoder_2"] = None
-                    extra_pipeline_kwargs["tokenizer_2"] = None
+        pipeline_type = (
+            PipelineTypes.CONTROLNET
+            if self.config.controlnet
+            else (
+                PipelineTypes.CONTROL
+                if self.config.control
+                else self.model.DEFAULT_PIPELINE_TYPE
+            )
+        )
+        self.model.pipeline = self.model.get_pipeline(
+            pipeline_type=pipeline_type,
+            load_base_model=False,
+        )
 
-            if self.args.model_family == "smoldit":
-                extra_pipeline_kwargs["transformer"] = unwrap_model(
-                    self.accelerator, self.transformer
-                )
-                extra_pipeline_kwargs["tokenizer"] = self.tokenizer_1
-                extra_pipeline_kwargs["text_encoder"] = self.text_encoder_1
-                extra_pipeline_kwargs["scheduler"] = self.setup_scheduler()
-
-            if self.args.controlnet:
-                # ControlNet training has an additional adapter thingy.
-                extra_pipeline_kwargs["controlnet"] = unwrap_model(
-                    self.accelerator, self.controlnet
-                )
-            if self.unet is not None:
-                extra_pipeline_kwargs["unet"] = unwrap_model(
-                    self.accelerator, self.unet
-                )
-
-            if self.transformer is not None:
-                extra_pipeline_kwargs["transformer"] = unwrap_model(
-                    self.accelerator, self.transformer
-                )
-
-            if self.args.model_family == "sd3" and self.args.train_text_encoder:
-                if self.text_encoder_1 is not None:
-                    extra_pipeline_kwargs["text_encoder"] = unwrap_model(
-                        self.accelerator, self.text_encoder_1
-                    )
-                    extra_pipeline_kwargs["tokenizer"] = self.tokenizer_1
-                if self.text_encoder_2 is not None:
-                    extra_pipeline_kwargs["text_encoder_2"] = unwrap_model(
-                        self.accelerator, self.text_encoder_2
-                    )
-                    extra_pipeline_kwargs["tokenizer_2"] = self.tokenizer_2
-                if self.text_encoder_3 is not None:
-                    extra_pipeline_kwargs["text_encoder_3"] = unwrap_model(
-                        self.accelerator, self.text_encoder_3
-                    )
-                    extra_pipeline_kwargs["tokenizer_3"] = self.tokenizer_3
-
-            if self.vae is None or not hasattr(self.vae, "device"):
-                extra_pipeline_kwargs["vae"] = self.init_vae()
-            else:
-                logger.info(f"Found VAE: {self.vae.config}")
-            if (
-                "vae" in extra_pipeline_kwargs
-                and extra_pipeline_kwargs.get("vae") is not None
-                and extra_pipeline_kwargs["vae"].device != self.inference_device
-            ):
-                extra_pipeline_kwargs["vae"] = extra_pipeline_kwargs["vae"].to(
-                    self.inference_device
-                )
-
-            pipeline_kwargs = {
-                "pretrained_model_name_or_path": self.args.pretrained_model_name_or_path,
-                "revision": self.args.revision,
-                "variant": self.args.variant,
-                "torch_dtype": self.weight_dtype,
-                **extra_pipeline_kwargs,
-            }
-            logger.debug(f"Initialising pipeline with kwargs: {pipeline_kwargs}")
-            attempt = 0
-            while attempt < 3:
-                attempt += 1
-                try:
-                    if self.args.model_family == "smoldit":
-                        self.pipeline = pipeline_cls(
-                            vae=self.vae,
-                            transformer=unwrap_model(
-                                self.accelerator, self.transformer
-                            ),
-                            tokenizer=self.tokenizer_1,
-                            text_encoder=self.text_encoder_1,
-                            scheduler=self.setup_scheduler(),
-                        )
-                    else:
-                        self.pipeline = pipeline_cls.from_pretrained(**pipeline_kwargs)
-                except Exception as e:
-                    import traceback
-
-                    logger.error(e)
-                    logger.error(traceback.format_exc())
-                    continue
-                break
-            if self.args.validation_torch_compile:
-                if self.deepspeed:
-                    logger.warning(
-                        "DeepSpeed does not support torch compile. Disabling. Set --validation_torch_compile=False to suppress this warning."
-                    )
-                elif self.args.lora_type.lower() == "lycoris":
-                    logger.warning(
-                        "LyCORIS does not support torch compile for validation due to graph compile breaks. Disabling. Set --validation_torch_compile=False to suppress this warning."
-                    )
-                else:
-                    if self.unet is not None and not is_compiled_module(self.unet):
-                        logger.warning(
-                            f"Compiling the UNet for validation ({self.args.validation_torch_compile})"
-                        )
-                        self.pipeline.unet = torch.compile(
-                            self.pipeline.unet,
-                            mode=self.args.validation_torch_compile_mode,
-                            fullgraph=False,
-                        )
-                    if self.transformer is not None and not is_compiled_module(
-                        self.transformer
-                    ):
-                        logger.warning(
-                            f"Compiling the transformer for validation ({self.args.validation_torch_compile})"
-                        )
-                        self.pipeline.transformer = torch.compile(
-                            self.pipeline.transformer,
-                            mode=self.args.validation_torch_compile_mode,
-                            fullgraph=False,
-                        )
-
-        self.pipeline = self.pipeline.to(self.inference_device)
-        self.pipeline.set_progress_bar_config(disable=True)
+        self.model.move_models(self.accelerator.device)
+        # Remove text encoders on 'meta' device to avoid move errors
+        for attr in ["text_encoder", "text_encoder_2", "text_encoder_3", "text_encoder_4"]:
+            te = getattr(self.model.pipeline, attr, None)
+            if getattr(te, "device", None) and te.device.type == "meta":
+                setattr(self.model.pipeline, attr, None)
+        self.model.pipeline.to(self.accelerator.device)
+        self.model.pipeline.set_progress_bar_config(disable=True)
 
     def clean_pipeline(self):
         """Remove the pipeline."""
         if hasattr(self.accelerator, "_lycoris_wrapped_network"):
             self.accelerator._lycoris_wrapped_network.set_multiplier(1.0)
-        if self.pipeline is not None:
-            del self.pipeline
-            self.pipeline = None
+        if self.model.pipeline is not None:
+            del self.model.pipeline
+            self.model.pipeline = None
 
     def process_prompts(self):
         """Processes each validation prompt and logs the result."""
         self.validation_prompt_dict = {}
         self.evaluation_result = None
         validation_images = {}
-        _content = zip(self.validation_shortnames, self.validation_prompts)
-        total_samples = (
-            len(self.validation_shortnames)
-            if self.validation_shortnames is not None
-            else 0
-        )
+        _content = self.validation_prompt_metadata["validation_prompts"]
+        total_samples = len(_content) if _content is not None else 0
         self.eval_scores = {}
         if self.validation_image_inputs:
             # Override the pipeline inputs to be entirely based upon the validation image inputs.
             _content = self.validation_image_inputs
+            if "DeepFloyd" in self.model.NAME:
+                resize_edge_length = 64
+                # Resize validation input to 64px area
+                _content = resize_validation_images(_content, resize_edge_length)
             total_samples = len(_content) if _content is not None else 0
-        for content in tqdm(
+
+        logger.debug(f"Processing content: {_content}")
+        idx = 0
+        for prompt in tqdm(
             _content if _content else [],
             desc="Processing validation prompts",
             total=total_samples,
@@ -1331,15 +987,21 @@ class Validation:
             position=1,
         ):
             validation_input_image = None
-            logger.debug(f"content: {content}")
-            if len(content) == 3:
-                shortname, prompt, validation_input_image = content
-            elif len(content) == 2:
-                shortname, prompt = content
+            if len(prompt) == 3 and isinstance(prompt[2], Image.Image):
+                # DeepFloyd stage II inputs.
+                shortname, prompt, validation_input_image = prompt
+            elif len(prompt) == 4 and isinstance(prompt[3], Image.Image):
+                (
+                    shortname,
+                    prompt,
+                    validation_input_image_path,
+                    validation_input_image,
+                ) = prompt
             else:
-                raise ValueError(
-                    f"Validation content is not in the correct format: {content}"
-                )
+                shortname = self.validation_prompt_metadata["validation_shortnames"][
+                    idx
+                ]
+            logger.debug(f"validation prompt (shortname={shortname}): '{prompt}'")
             self.validation_prompt_dict[shortname] = prompt
             logger.debug(f"Processing validation for prompt: {prompt}")
             (
@@ -1348,7 +1010,7 @@ class Validation:
                 ema_validation_images,
             ) = self.validate_prompt(prompt, shortname, validation_input_image)
             validation_images.update(stitched_validation_images)
-            if self.args.model_family in ["ltxvideo", "wan"]:
+            if isinstance(self.model, VideoModelFoundation):
                 self._save_videos(validation_images, shortname, prompt)
             else:
                 self._save_images(validation_images, shortname, prompt)
@@ -1356,10 +1018,14 @@ class Validation:
             self.validation_images = validation_images
             self.evaluation_result = self.evaluate_images(checkpoint_validation_images)
             self._log_validations_to_webhook(validation_images, shortname, prompt)
+            idx += 1
         try:
             self._log_validations_to_trackers(validation_images)
         except Exception as e:
             logger.error(f"Error logging validation images: {e}")
+            import traceback
+
+            logger.error(traceback.format_exc())
 
     def get_eval_result(self):
         return self.evaluation_result or {}
@@ -1384,12 +1050,12 @@ class Validation:
 
     def _validation_types(self):
         types = ["checkpoint"]
-        if self.args.use_ema:
+        if self.config.use_ema:
             # ema has different validations we can add or overwrite.
-            if self.args.ema_validation == "ema_only":
+            if self.config.ema_validation == "ema_only":
                 # then we do not sample the base ckpt being trained, only the EMA weights.
                 types = ["ema"]
-            if self.args.ema_validation == "comparison":
+            if self.config.ema_validation == "comparison":
                 # then we sample both.
                 types.append("ema")
 
@@ -1400,62 +1066,87 @@ class Validation:
     ):
         """Generate validation images for a single prompt."""
         # Placeholder for actual image generation and logging
-        logger.debug(f"Validating prompt: {prompt}")
+        logger.debug(f"Validating ({validation_shortname}) prompt: {prompt}")
         # benchmarked / stitched validation images
         stitched_validation_images = {}
         # untouched / un-stitched validation images
         checkpoint_validation_images = {}
         ema_validation_images = {}
+        benchmark_image = None
         for resolution in self.validation_resolutions:
             extra_validation_kwargs = {}
             if validation_input_image is not None:
                 extra_validation_kwargs["image"] = validation_input_image
                 if self.deepfloyd_stage2:
+                    # deepfloyd-if stage 2 requires 4x size declaration on the outputs.
+                    # the input is 64px edge len.
                     validation_resolution_width, validation_resolution_height = (
                         val * 4 for val in extra_validation_kwargs["image"].size
                     )
-                elif self.args.controlnet or self.args.validation_using_datasets:
-                    validation_resolution_width, validation_resolution_height = (
-                        extra_validation_kwargs["image"].size
+                elif (
+                    self.config.controlnet
+                    or self.config.validation_using_datasets
+                    or self.model.requires_conditioning_validation_inputs()
+                ):
+                    # for most conditioned models, we want the input size to remain.
+                    validation_image_edge_len = (
+                        self.model.validation_image_input_edge_length()
                     )
+                    if validation_image_edge_len is not None:
+                        # calculate the megapixels value (eg ~0.25 for 512px)
+                        validation_image_megapixels = (
+                            validation_image_edge_len**2
+                        ) / 1_000_000
+                        validation_resolution, _, validation_aspect_ratio = (
+                            MultiaspectImage.calculate_new_size_by_pixel_area(
+                                aspect_ratio=MultiaspectImage.calculate_image_aspect_ratio(
+                                    extra_validation_kwargs["image"]
+                                ),
+                                megapixels=validation_image_megapixels,
+                                original_size=extra_validation_kwargs["image"].size,
+                            )
+                        )
+                        extra_validation_kwargs["image"] = extra_validation_kwargs[
+                            "image"
+                        ].resize(validation_resolution, Image.Resampling.LANCZOS)
+                        validation_resolution_width, validation_resolution_height = (
+                            validation_resolution
+                        )
+                    else:
+                        validation_resolution_width, validation_resolution_height = (
+                            extra_validation_kwargs["image"].size
+                        )
                 else:
                     raise ValueError(
                         "Validation input images are not supported for this model type."
                     )
+                extra_validation_kwargs["control_image"] = extra_validation_kwargs[
+                    "image"
+                ]
             else:
                 validation_resolution_width, validation_resolution_height = resolution
 
-            if type(self.args.validation_guidance_skip_layers) is list:
+            if type(self.config.validation_guidance_skip_layers) is list:
                 extra_validation_kwargs["skip_layer_guidance_start"] = float(
-                    self.args.validation_guidance_skip_layers_start
+                    self.config.validation_guidance_skip_layers_start
                 )
                 extra_validation_kwargs["skip_layer_guidance_stop"] = float(
-                    self.args.validation_guidance_skip_layers_stop
+                    self.config.validation_guidance_skip_layers_stop
                 )
                 extra_validation_kwargs["skip_layer_guidance_scale"] = float(
-                    self.args.validation_guidance_skip_scale
+                    self.config.validation_guidance_skip_scale
                 )
                 extra_validation_kwargs["skip_guidance_layers"] = list(
-                    self.args.validation_guidance_skip_layers
+                    self.config.validation_guidance_skip_layers
                 )
 
-            if not self.flow_matching and self.args.model_family not in [
-                "deepfloyd",
-                "pixart_sigma",
-                "kolors",
-                "flux",
-                "sd3",
-                "sana",
-                "ltxvideo",
-                "wan",
-            ]:
-                extra_validation_kwargs["guidance_rescale"] = (
-                    self.args.validation_guidance_rescale
-                )
+            extra_validation_kwargs["guidance_rescale"] = (
+                self.config.validation_guidance_rescale
+            )
 
             if StateTracker.get_args().validation_using_datasets:
                 extra_validation_kwargs["strength"] = getattr(
-                    self.args, "validation_strength", 0.2
+                    self.config, "validation_strength", 0.2
                 )
                 logger.debug(
                     f"Set validation image denoise strength to {extra_validation_kwargs['strength']}"
@@ -1469,7 +1160,11 @@ class Validation:
                 checkpoint_validation_images[validation_shortname] = []
                 ema_validation_images[validation_shortname] = []
             try:
-                extra_validation_kwargs.update(self._gather_prompt_embeds(prompt))
+                _embed = self._gather_prompt_embeds(prompt)
+                if _embed is not None:
+                    extra_validation_kwargs.update(_embed)
+                else:
+                    extra_validation_kwargs["prompt"] = prompt
             except Exception as e:
                 import traceback
 
@@ -1482,9 +1177,9 @@ class Validation:
                 pipeline_kwargs = {
                     "prompt": None,
                     "negative_prompt": None,
-                    "num_images_per_prompt": self.args.num_validation_images,
-                    "num_inference_steps": self.args.validation_num_inference_steps,
-                    "guidance_scale": self.args.validation_guidance,
+                    "num_images_per_prompt": self.config.num_validation_images,
+                    "num_inference_steps": self.config.validation_num_inference_steps,
+                    "guidance_scale": self.config.validation_guidance,
                     "height": MultiaspectImage._round_to_nearest_multiple(
                         int(validation_resolution_height), 16
                     ),
@@ -1493,82 +1188,123 @@ class Validation:
                     ),
                     **extra_validation_kwargs,
                 }
-                if self.args.validation_guidance_real > 1.0:
+                if self.model.VALIDATION_USES_NEGATIVE_PROMPT:
+                    if StateTracker.get_args().validation_negative_prompt is None:
+                        StateTracker.get_args().validation_negative_prompt = ""
+                    _negative_embed = self.embed_cache.compute_embeddings_for_prompts(
+                        [StateTracker.get_args().validation_negative_prompt],
+                        is_validation=True,
+                        load_from_cache=True,
+                    )
+                    if _negative_embed is not None:
+                        negative_embed_data = {
+                            k: (
+                                v.to(
+                                    device=self.inference_device,
+                                    dtype=self.config.weight_dtype,
+                                )
+                                if hasattr(v, "to")
+                                else v
+                            )
+                            for k, v in _negative_embed.items()
+                        }
+                        pipeline_kwargs.update(
+                            self.model.convert_negative_text_embed_for_pipeline(
+                                prompt=StateTracker.get_args().validation_negative_prompt,
+                                text_embedding=negative_embed_data,
+                            )
+                        )
+                    else:
+                        pipeline_kwargs["negative_prompt"] = (
+                            StateTracker.get_args().validation_negative_prompt
+                        )
+                # TODO: Refactor the rest so that it uses model class to update kwargs more generally.
+                if self.config.validation_guidance_real > 1.0:
                     pipeline_kwargs["guidance_scale_real"] = float(
-                        self.args.validation_guidance_real
+                        self.config.validation_guidance_real
                     )
                 if (
-                    isinstance(self.args.validation_no_cfg_until_timestep, int)
-                    and self.args.model_family == "flux"
+                    isinstance(self.config.validation_no_cfg_until_timestep, int)
+                    and self.config.model_family == "flux"
                 ):
                     pipeline_kwargs["no_cfg_until_timestep"] = (
-                        self.args.validation_no_cfg_until_timestep
+                        self.config.validation_no_cfg_until_timestep
                     )
 
+                pipeline_kwargs = self.model.update_pipeline_call_kwargs(
+                    pipeline_kwargs
+                )
                 logger.debug(
                     f"Image being generated with parameters: {pipeline_kwargs}"
                 )
                 # Print the device attr of any parameters that have one
-                for key, value in pipeline_kwargs.items():
-                    if hasattr(value, "device"):
-                        logger.debug(f"Device for {key}: {value.device}")
-                for key, value in self.pipeline.components.items():
-                    if hasattr(value, "device"):
-                        logger.debug(f"Device for {key}: {value.device}")
-                if StateTracker.get_model_family() == "flux":
-                    if "negative_prompt" in pipeline_kwargs:
-                        del pipeline_kwargs["negative_prompt"]
-                if StateTracker.get_model_family() in ["ltxvideo", "wan"]:
-                    del pipeline_kwargs["num_images_per_prompt"]
-                if self.args.model_family == "sana":
+                # for key, value in pipeline_kwargs.items():
+                #     if hasattr(value, "device"):
+                #         logger.debug(f"Device for {key}: {value.device}")
+                # for key, value in self.model.pipeline.components.items():
+                #     if hasattr(value, "device"):
+                #         logger.debug(f"Device for {key}: {value.device}")
+                if self.config.model_family == "sana":
                     pipeline_kwargs["complex_human_instruction"] = (
-                        self.args.sana_complex_human_instruction
+                        self.config.sana_complex_human_instruction
                     )
-                if StateTracker.get_model_family() in [
-                    "wan",
-                ]:
-                    # Wan video should max out around 81 frames for efficiency.
-                    pipeline_kwargs["num_frames"] = min(
-                        81, self.args.validation_num_video_frames or 81
-                    )
-                    pipeline_kwargs["output_type"] = "pil"
-                    # replace embeds with prompt
-
-                if StateTracker.get_model_family() in [
-                    "pixart_sigma",
-                    "smoldit",
-                    "sana",
-                    "ltxvideo",
-                ]:
-                    if pipeline_kwargs.get("negative_prompt") is not None:
-                        del pipeline_kwargs["negative_prompt"]
-                    if pipeline_kwargs.get("prompt") is not None:
-                        del pipeline_kwargs["prompt"]
-                    pipeline_kwargs["prompt_attention_mask"] = pipeline_kwargs.pop(
-                        "prompt_mask"
-                    )[0].to(device=self.inference_device, dtype=self.weight_dtype)
-                    pipeline_kwargs["negative_prompt_attention_mask"] = torch.unsqueeze(
-                        pipeline_kwargs.pop("negative_mask")[0], dim=0
-                    ).to(device=self.inference_device, dtype=self.weight_dtype)
 
                 validation_types = self._validation_types()
                 all_validation_type_results = {}
                 for current_validation_type in validation_types:
-                    if not self.args.validation_randomize:
+                    if not self.config.validation_randomize:
                         pipeline_kwargs["generator"] = self._get_generator()
                         logger.debug(
                             f"Using a generator? {pipeline_kwargs['generator']}"
                         )
                     if current_validation_type == "ema":
                         self.enable_ema_for_inference()
-                    if self.args.model_family in ["ltxvideo", "wan"]:
-                        all_validation_type_results[current_validation_type] = (
-                            self.pipeline(**pipeline_kwargs).frames
+                    pipeline_kwargs = {
+                        k: (
+                            v.to(
+                                device=self.inference_device,
+                                dtype=self.config.weight_dtype,
+                            )
+                            if hasattr(v, "to")
+                            else v
                         )
-                    else:
-                        all_validation_type_results[current_validation_type] = (
-                            self.pipeline(**pipeline_kwargs).images
+                        for k, v in pipeline_kwargs.items()
+                    }
+                    import inspect
+
+                    call_kwargs = inspect.signature(
+                        self.model.pipeline.__call__
+                    ).parameters
+                    logger.debug(
+                        f"Possible parameters for {type(self.model.pipeline)}: {call_kwargs}"
+                    )
+                    # remove any kwargs that are not in the pipeline call
+                    pipeline_kwargs = {
+                        k: v for k, v in pipeline_kwargs.items() if k in call_kwargs
+                    }
+                    removed_kwargs = [
+                        k for k in pipeline_kwargs.keys() if k not in call_kwargs
+                    ]
+                    logger.debug(
+                        f"Running validations with inputs: {pipeline_kwargs.keys()}"
+                    )
+                    if removed_kwargs:
+                        logger.warning(
+                            f"Removed the following kwargs from validation pipeline: {removed_kwargs}"
                         )
+                    # run in autocast ctx
+                    with torch.amp.autocast(
+                        self.inference_device.type,
+                        dtype=self.config.weight_dtype,
+                    ):
+                        if isinstance(self.model, VideoModelFoundation):
+                            all_validation_type_results[current_validation_type] = (
+                                self.model.pipeline(**pipeline_kwargs).frames
+                            )
+                        elif isinstance(self.model, ImageModelFoundation):
+                            all_validation_type_results[current_validation_type] = (
+                                self.model.pipeline(**pipeline_kwargs).images
+                            )
                     if current_validation_type == "ema":
                         self.disable_ema_for_inference()
 
@@ -1579,12 +1315,12 @@ class Validation:
                 )
                 original_validation_image_results = validation_image_results
                 benchmark_image = None
-                if self.args.controlnet:
+                if any([self.config.controlnet, self.config.control]):
                     validation_image_results = self.stitch_conditioning_images(
                         original_validation_image_results,
-                        extra_validation_kwargs["image"],
+                        extra_validation_kwargs["control_image"],
                     )
-                elif not self.args.disable_benchmark and self.benchmark_exists(
+                elif not self.config.disable_benchmark and self.benchmark_exists(
                     "base_model"
                 ):
                     benchmark_image = self._benchmark_image(
@@ -1605,10 +1341,21 @@ class Validation:
                 stitched_validation_images[validation_shortname].extend(
                     validation_image_results
                 )
-                if self.args.use_ema:
-                    ema_validation_images[validation_shortname].extend(
-                        ema_image_results
-                    )
+                if self.config.use_ema:
+                    if (
+                        validation_shortname in ema_validation_images
+                        and ema_image_results is not None
+                    ):
+                        if ema_validation_images[validation_shortname] is None:
+                            # init the value
+                            ema_validation_images[validation_shortname] = []
+                        if isinstance(
+                            ema_validation_images[validation_shortname], list
+                        ):
+                            # if we have a list of images, we can stitch them.
+                            ema_validation_images[validation_shortname].extend(
+                                ema_image_results
+                            )
 
             except Exception as e:
                 import traceback
@@ -1618,8 +1365,8 @@ class Validation:
                 )
                 continue
         if (
-            self.args.use_ema
-            and self.args.ema_validation == "comparison"
+            self.config.use_ema
+            and self.config.ema_validation == "comparison"
             and benchmark_image is not None
         ):
             for idx, validation_image in enumerate(
@@ -1658,7 +1405,7 @@ class Validation:
                     self.save_dir,
                     f"step_{StateTracker.get_global_step()}_{validation_shortname}_{validation_img_idx}_{res_label}.mp4",
                 ),
-                fps=self.args.framerate,
+                fps=self.config.framerate,
             )
             validation_img_idx += 1
 
@@ -1722,7 +1469,7 @@ class Validation:
                     f"{res[0]}x{res[1]}" for res in get_validation_resolutions()
                 ]
 
-                if self.args.tracker_image_layout == "table":
+                if self.config.tracker_image_layout == "table":
                     columns = [
                         "Prompt",
                         *resolution_list,
@@ -1756,13 +1503,18 @@ class Validation:
                         step=StateTracker.get_global_step(),
                     )
 
-                elif self.args.tracker_image_layout == "gallery":
+                elif self.config.tracker_image_layout == "gallery":
                     gallery_images = {}
                     for prompt_shortname, image_list in validation_images.items():
                         logger.debug(
                             f"Prompt {prompt_shortname} has {len(image_list)} images"
                         )
                         for idx, image in enumerate(image_list):
+                            # if it's a list of images, make a grid
+                            if isinstance(image, list) and isinstance(
+                                image[0], Image.Image
+                            ):
+                                image = image[0]
                             wandb_image = wandb.Image(
                                 image,
                                 caption=f"{prompt_shortname} - {resolution_list[idx]}",
@@ -1774,23 +1526,15 @@ class Validation:
                     # Log all images in one call to prevent the global step from ticking
                     tracker.log(gallery_images, step=StateTracker.get_global_step())
 
-    def _primary_model(self):
-        if self.args.controlnet:
-            return self.controlnet
-        if self.unet is not None:
-            return self.unet
-        if self.transformer is not None:
-            return self.transformer
-
     def enable_ema_for_inference(self, pipeline=None):
         if self.ema_enabled:
             logger.debug("EMA already enabled. Not enabling EMA.")
             return
-        if self.args.use_ema:
+        if self.config.use_ema:
             logger.debug("Enabling EMA.")
             self.ema_enabled = True
-            if self.args.model_type == "lora":
-                if self.args.lora_type.lower() == "lycoris":
+            if self.config.model_type == "lora":
+                if self.config.lora_type.lower() == "lycoris":
                     logger.debug("Setting Lycoris multiplier to 1.0")
                     self.accelerator._lycoris_wrapped_network.set_multiplier(1.0)
                     logger.debug("Storing Lycoris weights for later recovery.")
@@ -1803,23 +1547,25 @@ class Validation:
                     self.ema_model.copy_to(
                         self.accelerator._lycoris_wrapped_network.parameters()
                     )
-                elif self.args.lora_type.lower() == "standard":
+                elif self.config.lora_type.lower() == "standard":
                     _trainable_parameters = [
-                        x for x in self._primary_model().parameters() if x.requires_grad
+                        x
+                        for x in self.model.get_trained_component().parameters()
+                        if x.requires_grad
                     ]
                     self.ema_model.store(_trainable_parameters)
                     self.ema_model.copy_to(_trainable_parameters)
             else:
-                # if self.args.ema_device != "accelerator":
+                # if self.config.ema_device != "accelerator":
                 #     logger.info("Moving checkpoint to CPU for storage.")
-                #     self._primary_model().to("cpu")
+                #     self.model.get_trained_component().to("cpu")
                 logger.debug("Storing EMA weights for later recovery.")
                 self.ema_model.store(self.trainable_parameters())
                 logger.debug("Storing the EMA weights into the model for inference.")
                 self.ema_model.copy_to(self.trainable_parameters())
-            # if self.args.ema_device != "accelerator":
+            # if self.config.ema_device != "accelerator":
             #     logger.debug("Moving checkpoint to CPU for storage.")
-            #     self._primary_model().to("cpu")
+            #     self.model.get_trained_component().to("cpu")
             #     logger.debug("Moving EMA weights to GPU for inference.")
             #     self.ema_model.to(self.inference_device)
         else:
@@ -1831,12 +1577,12 @@ class Validation:
         if not self.ema_enabled:
             logger.debug("EMA was not enabled. Not disabling EMA.")
             return
-        if self.args.use_ema:
+        if self.config.use_ema:
             logger.debug("Disabling EMA.")
             self.ema_enabled = False
             if (
-                self.args.model_type == "lora"
-                and self.args.lora_type.lower() == "lycoris"
+                self.config.model_type == "lora"
+                and self.config.lora_type.lower() == "lycoris"
             ):
                 logger.debug("Setting Lycoris network multiplier to 1.0.")
                 self.accelerator._lycoris_wrapped_network.set_multiplier(1.0)
@@ -1847,10 +1593,10 @@ class Validation:
             else:
                 logger.debug("Restoring trainable parameters.")
                 self.ema_model.restore(self.trainable_parameters())
-            if self.args.ema_device != "accelerator":
+            if self.config.ema_device != "accelerator":
                 logger.debug("Moving EMA weights to CPU for storage.")
-                self.ema_model.to(self.args.ema_device)
-                self._primary_model().to(self.inference_device)
+                self.ema_model.to(self.config.ema_device)
+                self.model.get_trained_component().to(self.inference_device)
 
         else:
             logger.debug(
@@ -1859,10 +1605,9 @@ class Validation:
 
     def finalize_validation(self, validation_type):
         """Cleans up and restores original state if necessary."""
-        if not self.args.keep_vae_loaded and not self.args.vae_cache_ondemand:
-            self.vae = self.vae.to("cpu")
-            self.vae = None
-        self.pipeline = None
+        if not self.config.keep_vae_loaded and not self.config.vae_cache_ondemand:
+            self.model.unload_vae()
+        self.model.pipeline = None
         if torch.cuda.is_available():
             torch.cuda.empty_cache()
 
@@ -1876,6 +1621,11 @@ class Validation:
             for image in image_list:
                 evaluation_score = self.model_evaluator.evaluate([image], [prompt])
                 self.eval_scores[shortname] = round(float(evaluation_score), 4)
+        if len(self.eval_scores) == 0:
+            logger.warning(
+                "No evaluation scores were calculated. Please check your evaluation settings."
+            )
+            return {}
         # Log the scores into dict: {"min", "max", "mean", "std"}
         result = {
             "clip/min": min(self.eval_scores.values()),
@@ -1924,7 +1674,18 @@ class Evaluation:
         return sum(len(x["sampler"]) for _, x in eval_datasets.items())
 
     def get_timestep_schedule(self, noise_scheduler):
-        noise_scheduler.set_timesteps(self.config.eval_timesteps)
+        accept_mu = "mu" in set(
+            inspect.signature(noise_scheduler.set_timesteps).parameters.keys()
+        )
+        scheduler_kwargs = {}
+        if accept_mu and self.config.flow_schedule_auto_shift:
+            from helpers.models.sd3.pipeline import calculate_shift
+
+            scheduler_kwargs["mu"] = calculate_shift(
+                StateTracker.get_model().get_trained_component().config.max_seq
+            )
+
+        noise_scheduler.set_timesteps(self.config.eval_timesteps, **scheduler_kwargs)
         timesteps = noise_scheduler.timesteps
         return timesteps
 
@@ -2030,8 +1791,7 @@ class Evaluation:
                         )
                         eval_loss = calculate_loss(
                             prepared_batch=prepared_eval_batch,
-                            model_pred=eval_prediction,
-                            target=get_prediction_target(prepared_eval_batch),
+                            model_output=eval_prediction,
                             apply_conditioning_mask=False,
                         )
                         accumulated_eval_losses[eval_timestep].append(eval_loss)
@@ -2070,10 +1830,12 @@ class Evaluation:
         eval_datasets = StateTracker.get_data_backends(
             _type="eval", _types=None
         )  # dict of {name: ...}
+        if len(eval_datasets) == 0:
+            return {}
 
         if pooling:
             # Single pass across ALL eval datasets
-            logger.info("Running a single pooled eval pass across all datasets.")
+            logger.debug("Running a single pooled eval pass across all datasets.")
             pooled_losses = self._evaluate_dataset_pass(
                 dataset_name=None,
                 prepare_batch=prepare_batch,
@@ -2087,7 +1849,7 @@ class Evaluation:
 
         else:
             # Multiple passes: one per dataset
-            logger.info(
+            logger.debug(
                 "Running separate eval passes for each dataset + pooled results."
             )
             accumulated = {}
@@ -2128,6 +1890,8 @@ class Evaluation:
         """
         if not self.accelerator.is_main_process:
             return {}
+        if all_accumulated_losses == {} or all_accumulated_losses is None:
+            return {}
 
         results = {}
 
@@ -2139,7 +1903,7 @@ class Evaluation:
                     data_rows.append((ts, loss_val))
             return data_rows
 
-        logger.info("Generating evaluation tracker tables...")
+        logger.debug("Generating evaluation tracker tables...")
         for ds_name, timestep_dict in all_accumulated_losses.items():
             data_rows = flatten_timestep_losses(timestep_dict)
             if not data_rows:
