@@ -9,7 +9,6 @@ from helpers.prompts import (
 class TestPromptHandler(unittest.TestCase):
     def setUp(self):
         self.args = MagicMock()
-        self.args.disable_compel = False
         self.text_encoders = [MagicMock(), MagicMock()]
         self.tokenizers = [MagicMock(), MagicMock()]
         self.accelerator = MagicMock()
@@ -104,6 +103,7 @@ class TestPromptHandler(unittest.TestCase):
             text_encoders=["LameTest"],
             tokenizers=["LameTest"],
             accelerator=MagicMock(),
+            model_type="sdxl",
         )
         result_caption = handler.magic_prompt(
             "path/to/image.png",
@@ -129,6 +129,7 @@ class TestPromptHandler(unittest.TestCase):
             text_encoders=["LameTest"],
             tokenizers=["LameTest"],
             accelerator=MagicMock(),
+            model_type="sdxl",
         )
         result_caption = handler.magic_prompt(
             f"path/to/{image_filename}.png",
